feat(mobile): Warm Editorial fonts app-wide + Learn article extraction
Learn step 3: extract the 7 inline articles (strokes-gained, benchmarks, glossary, how-to-practice, course-management, mental-game, skill-games-pressure-games) from the 2,230-line [article].tsx onto the shared learn primitives; add a dev-only DevNote primitive (preserves __DEV__-gated source/TODO notes); collapse [article].tsx to a 99-line router; centralize the learn palette.

Full font adoption: add role-named TYPE presets to lib/typography and apply them to every <Text> across the app (home, rounds, live round, hole entry, stats, patterns, practice, bag, profile, auth, onboarding, errors). Typeface-only — sizes/colors/spacing preserved. Centralizes all previously-hardcoded family strings and fixes stale Inter refs that were silently falling back to system.

Typecheck green. Device/simulator QA pending (SVG/map labels, share cards).
