+++ apps/mobile/app/(app)/learn/[article].tsx
@@ -1,8 +1,8 @@
-import { Linking, Pressable, ScrollView, Text, View } from 'react-native'
+import { Pressable, ScrollView, Text, View } from 'react-native'
 import { useLocalSearchParams, useRouter } from 'expo-router'
 import { findLearnArticle } from '@oga/core'
 import { AppBar } from '../../../components/ui/AppBar'
-import { BODY, KICKER, SUBKICKER, TITLE } from '../../../components/learn/primitives'
+import { BODY, C, KICKER, TITLE } from '../../../components/learn/primitives'
 import { MOBILE_ARTICLES } from '../../../components/learn/articles'
 
 export default function ArticleScreen() {
@@ -11,7 +11,7 @@ export default function ArticleScreen() {
   const found = slug ? findLearnArticle(slug) : null
 
   return (
-    <View style={{ flex: 1, backgroundColor: '#F2EEE5' }}>
+    <View style={{ flex: 1, backgroundColor: C.bg }}>
       <AppBar
         eyebrow={found?.section.title ?? 'Yardage book'}
         title={found?.article.title ?? 'Article'}
@@ -39,11 +39,18 @@ export default function ArticleScreen() {
   )
 }
 
+/** Render the per-article component registered for this catalog id. */
+function LiveArticle({ id }: { id: string }) {
+  const Article = MOBILE_ARTICLES[id]
+  if (Article) return <Article />
+  return <StubBody title="Article" />
+}
+
 function NotFound() {
   return (
     <View>
       <Text style={TITLE}>Article not found.</Text>
-      <Text style={{ ...BODY, color: '#5C6356', fontStyle: 'italic' }}>
+      <Text style={{ ...BODY, color: C.inkDim, fontStyle: 'italic' }}>
         That guide does not exist yet.
       </Text>
     </View>
@@ -55,23 +62,23 @@ function DraftBanner() {
     <View
       style={{
         borderLeftWidth: 3,
-        borderLeftColor: '#A66A1F',
+        borderLeftColor: C.amber,
         borderTopWidth: 1,
         borderRightWidth: 1,
         borderBottomWidth: 1,
-        borderTopColor: '#D9D2BF',
-        borderRightColor: '#D9D2BF',
-        borderBottomColor: '#D9D2BF',
+        borderTopColor: C.line,
+        borderRightColor: C.line,
+        borderBottomColor: C.line,
         borderRadius: 2,
-        backgroundColor: '#FBF8F1',
+        backgroundColor: C.surface,
         padding: 14,
         marginBottom: 18,
       }}
     >
-      <Text style={{ ...KICKER, color: '#A66A1F', marginBottom: 6 }}>
+      <Text style={{ ...KICKER, color: C.amber, marginBottom: 6 }}>
         Work in progress
       </Text>
-      <Text style={{ color: '#1C211C', fontSize: 13, lineHeight: 19 }}>
+      <Text style={{ color: C.ink, fontSize: 13, lineHeight: 19 }}>
         This guide is being reviewed for accuracy. Treat specific technique
         advice as provisional until the notice is removed.
       </Text>
@@ -84,2147 +91,9 @@ function StubBody({ title }: { title: string }) {
     <View>
       <Text style={{ ...KICKER, marginBottom: 10 }}>Coming soon</Text>
       <Text style={TITLE}>{title}</Text>
-      <Text style={{ ...BODY, color: '#8A8B7E', fontStyle: 'italic' }}>
+      <Text style={{ ...BODY, color: C.mute, fontStyle: 'italic' }}>
         This guide is being written. Check back soon.
       </Text>
     </View>
   )
 }
-
-function LiveArticle({ id }: { id: string }) {
-  switch (id) {
-    case 'strokes-gained':
-      return <StrokesGainedArticle />
-    case 'benchmarks':
-      return <BenchmarksArticle />
-    case 'glossary':
-      return <GlossaryArticle />
-    case 'how-to-practice':
-      return <HowToPracticeArticle />
-    case 'course-management':
-      return <CourseManagementArticle />
-    case 'mental-game':
-      return <MentalGameArticle />
-    case 'skill-games-pressure-games':
-      return <SkillGamesArticle />
-  }
-  const Ported = MOBILE_ARTICLES[id]
-  if (Ported) return <Ported />
-  return <StubBody title="Article" />
-}
-
-function StrokesGainedArticle() {
-  return (
-    <View>
-      <Text style={{ ...KICKER, marginBottom: 10 }}>Strokes gained</Text>
-      <Text style={TITLE}>Where strokes come from.</Text>
-      <Text style={BODY}>
-        Every shot is graded against an expected outcome at your handicap. Beat
-        the expectation, you gain strokes; come up short, you lose them. Sum
-        across a round and you find out which part of the game is paying you
-        and which is leaking. Score alone tells you the result; SG tells you
-        why.
-      </Text>
-
-      <Text style={SUBKICKER}>The four categories</Text>
-      <Text style={BODY}>
-        Off tee = par-4 and par-5 tee shots. Approach = anything outside 30 yd
-        that is not a tee shot. Around the green = inside 30 yd, not on the
-        green. Putting = every shot on the green.
-      </Text>
-    </View>
-  )
-}
-
-function BenchmarksArticle() {
-  return (
-    <View>
-      <Text style={{ ...KICKER, marginBottom: 10 }}>By the numbers</Text>
-      <Text style={TITLE}>Where the field sits.</Text>
-      <Text style={BODY}>
-        Reference benchmarks for the stats this app tracks, from a 25-handicap
-        weekend round up to the PGA Tour. Use the table to see where you sit
-        relative to the bracket above and below your handicap.
-      </Text>
-
-      {BENCHMARKS.map((b) => (
-        <View
-          key={b.label}
-          style={{
-            borderWidth: 1,
-            borderColor: '#D9D2BF',
-            backgroundColor: '#FBF8F1',
-            borderRadius: 4,
-            padding: 14,
-            marginBottom: 10,
-          }}
-        >
-          <Text style={{ ...KICKER, marginBottom: 8 }}>{b.label}</Text>
-          <View style={{ flexDirection: 'row', flexWrap: 'wrap', gap: 14 }}>
-            <Stat label="PGA" value={b.pga} />
-            <Stat label="Scratch" value={b.scratch} />
-            <Stat label="Mid" value={b.mid} />
-            <Stat label="High" value={b.high} />
-          </View>
-        </View>
-      ))}
-
-      <Text style={{ ...KICKER, color: '#8A8B7E', marginTop: 14, lineHeight: 14 }}>
-        Benchmarks based on Mark Broadie's strokes gained research and PGA
-        Tour ShotLink data. Amateur averages approximate.
-      </Text>
-    </View>
-  )
-}
-
-function GlossaryArticle() {
-  return (
-    <View>
-      <Text style={{ ...KICKER, marginBottom: 10 }}>Understanding the game · Draft</Text>
-      <Text style={TITLE}>Glossary of golf terms.</Text>
-      <Para>
-        The basics — birdie, bogey, par, handicap, GIR — are covered
-        in the app's stat reference section. This glossary focuses
-        on the terms that come up once you start playing regularly
-        and trying to improve.
-      </Para>
-
-      {GLOSSARY_GROUPS.map((group) => (
-        <View key={group.heading}>
-          <Divider />
-          <H3>{group.heading}</H3>
-          {group.intro && <Para>{group.intro}</Para>}
-          {group.terms.map((t) => (
-            <Term key={t.name} name={t.name}>
-              {t.body}
-            </Term>
-          ))}
-        </View>
-      ))}
-
-      <View
-        style={{
-          borderTopWidth: 1,
-          borderColor: '#D9D2BF',
-          paddingTop: 18,
-          marginTop: 22,
-        }}
-      >
-        <Text style={{ ...KICKER, color: '#8A8B7E', lineHeight: 14 }}>
-          Last reviewed May 2026 · Draft, being expanded · Edit
-          docs/learn/glossary.md to contribute
-        </Text>
-      </View>
-    </View>
-  )
-}
-
-function Term({ name, children }: { name: string; children: React.ReactNode }) {
-  return (
-    <View
-      style={{
-        borderTopWidth: 1,
-        borderColor: '#D9D2BF',
-        paddingVertical: 12,
-      }}
-    >
-      <Text
-        style={{
-          color: '#1C211C',
-          fontSize: 16,
-          fontStyle: 'italic',
-          fontWeight: '500',
-          marginBottom: 4,
-        }}
-      >
-        {name}
-      </Text>
-      <Text style={{ color: '#1C211C', fontSize: 14, lineHeight: 20 }}>
-        {children}
-      </Text>
-    </View>
-  )
-}
-
-interface GlossaryTerm {
-  name: string
-  body: string
-}
-
-interface GlossaryGroup {
-  heading: string
-  intro?: string
-  terms: GlossaryTerm[]
-}
-
-const GLOSSARY_GROUPS: GlossaryGroup[] = [
-  {
-    heading: 'Shot shapes and ball flight',
-    terms: [
-      { name: 'Draw', body: 'A controlled shot that curves gently from right to left for a right-handed golfer. Typically more roll than a fade. Not to be confused with a hook.' },
-      { name: 'Fade', body: 'A controlled shot that curves gently from left to right for a right-handed golfer. Tour pros often prefer it because the ball lands softer with less roll.' },
-      { name: 'Hook', body: 'An uncontrolled, exaggerated draw. Often the result of a closed clubface at impact.' },
-      { name: 'Slice', body: 'An uncontrolled, exaggerated fade. The most common miss for amateurs — open face plus out-to-in path.' },
-      { name: 'Push', body: 'A shot that flies straight but to the right of the target with no curve. Different from a fade — wrong line, no shape.' },
-      { name: 'Pull', body: 'A shot that flies straight but to the left of the target with no curve. Different from a hook.' },
-      { name: 'Stinger', body: 'A low, penetrating shot hit with minimal loft, played into wind or under trees. Made famous by Tiger Woods.' },
-      { name: 'Punch shot', body: 'A low, controlled shot played with a shortened swing. Stays low, runs out after landing.' },
-      { name: 'Flier', body: 'A shot that travels farther than expected because grass gets between face and ball, reducing backspin. Common from rough.' },
-      { name: 'Spinner', body: 'A wedge struck cleanly into a soft green that spins back toward the player after landing. Useful when the pin is at the back of the green.' },
-      { name: 'Knuckleball', body: 'A shot with almost no spin — the ball flutters unpredictably. Often happens out of thick rough.' },
-      { name: 'Texas Wedge', body: 'Using a putter from off the green — fringe, apron, or even light rough. Highest-percentage shot when the ground between ball and hole is firm and flat.' },
-      { name: 'DOD (Driver Off the Deck)', body: 'Hitting a driver without a tee, from the fairway or rough. One of the hardest shots in golf. Made famous by social media creator "DOD King" Carter Smith.' },
-      { name: 'Thai Spinner', body: 'A low, high-spin shot popularized by Kiradech Aphibarnrat — steep outside-in lob wedge from a tight or grainy lie. Reached wider audiences when Keith Mitchell pulled it off at the 2025 Houston Open.' },
-    ],
-  },
-  {
-    heading: 'Short game shots',
-    terms: [
-      { name: 'Chip', body: 'Low, running shot from just off the green. Less-lofted club, putting-style stroke. Land just on, let it roll.' },
-      { name: 'Pitch', body: 'Higher, softer shot from 20–80 yards. Wedge, fuller swing than a chip. Harder because partial swings demand distance control.' },
-      { name: 'Bump and run', body: 'Mid-iron chip that lands short of the green and runs onto the surface. Favored in firm conditions. Lower risk than a flop.' },
-      { name: 'Flop shot', body: 'High, soft lob-wedge shot that lands softly with minimal roll. High risk — easy to skull or chunk if not practiced.' },
-      { name: 'Chunk / Fat shot', body: 'Hitting the ground behind the ball. Club digs, decelerates, ball comes up well short.' },
-      { name: 'Scalp / Thin / Blade', body: 'Hitting the ball with the leading edge instead of the face. Comes out low and hot, runs through the green.' },
-      { name: 'Flush', body: 'Slang for a perfectly struck shot — clean contact in the center of the face.' },
-    ],
-  },
-  {
-    heading: 'Putting terms',
-    terms: [
-      { name: 'Pin high / Hole high', body: 'Approach finishes level with the pin (correct distance, off line). Generally a good miss — putting across slope, not up or down it.' },
-      { name: 'Spinner (putting)', body: 'A putt that lips out spinning — catches the edge, spins around the rim, comes back out. See also: lip out.' },
-    ],
-  },
-  {
-    heading: 'Course and green terminology',
-    terms: [
-      { name: 'Below the hole', body: 'The lower side of the cup on a sloped green. Leaves an uphill putt — preferred. Tour players always try to stay below.' },
-      { name: 'Above the hole', body: 'The upper side. Leaves a downhill putt — faster, harder to control. Avoid.' },
-      { name: 'Short side', body: 'The side of the green closest to a tucked pin. Missing short side leaves the hardest possible chip. Do not short side yourself.' },
-      { name: 'Grain', body: 'Direction grass grows on a green. Bermuda grain strongly affects speed and break. With the grain is faster; against is slower.' },
-      { name: 'Stimp / Stimpmeter', body: 'Device that measures green speed. 10 = club golf, 12–13 = major championship, 14+ = extremely fast.' },
-      { name: 'False front', body: 'A green that slopes back toward the fairway at the front. Approach shots that land on it roll back off. Take more club.' },
-      { name: 'Apron / Fringe', body: 'Short grass surrounding the green — shorter than fairway, longer than green. Putt-able with care.' },
-      { name: 'Dogleg', body: 'A hole that bends left or right between tee and green.' },
-      { name: 'Bailout', body: 'A safe area to miss to when the primary target is dangerous. Identify before swinging.' },
-      { name: 'Big ball (Earth)', body: 'Slang for chunking — hitting the ground before the ball. The "big ball" is the Earth; you want to hit the small one first.' },
-    ],
-  },
-  {
-    heading: 'Divots and course care',
-    terms: [
-      { name: 'Divot', body: 'The chunk of turf displaced (or the hole left). Replace it, or fill with sand-mix where provided. Hitting from a divot is no relief — play it as it lies.' },
-      { name: 'Pitch mark / Ball mark', body: 'The indentation an approach leaves on the green. Repair yours and any nearby. Push edges down and smooth — do not lift the center.' },
-      { name: 'Burn mark', body: 'Browning on one side of the cup where grain grows away from the cup edge. The healthy side is with the grain; the burned side is into it. Quick read for grain on Bermuda greens.' },
-      { name: 'Divot tool / Pitch fork', body: 'Tool for repairing pitch marks. Insert tines at the edges, push inward. Never pry up from the center.' },
-    ],
-  },
-  {
-    heading: 'Scoring and competition',
-    terms: [
-      { name: 'Snowman', body: 'A score of 8 on a hole — the number 8 looks like a snowman.' },
-      { name: 'Blow up hole', body: 'A hole where everything goes wrong and a large number is recorded. Mental game = minimizing how many.' },
-      { name: 'Honors', body: 'The right to tee off first. In stroke play, goes to the lowest score on the previous hole.' },
-      { name: 'Ready golf', body: 'Hit when you are ready, not strictly by honors order. Speeds pace of play. Standard in casual rounds.' },
-      { name: 'Skins', body: 'Each hole worth a fixed amount. Win the hole outright (no ties) and you win the skin. Ties carry over and stack.' },
-      { name: 'Match play', body: 'Score tracked hole by hole rather than total strokes. Win a hole, go one up.' },
-      { name: 'Stroke play', body: 'Total strokes over the round. Lowest total wins.' },
-      { name: 'Stableford', body: 'Points based on score relative to par. Bogey 1, par 2, birdie 3, eagle 4. Encourages aggressive play.' },
-      { name: 'Nassau', body: 'A bet split into front nine, back nine, and full 18 — three separate match-play competitions.' },
-      { name: 'Press', body: 'A new side bet started mid-round in a Nassau or match play, typically when down by two.' },
-      { name: 'Dormie', body: 'In match play, leading by the same number of holes as remain. Cannot lose — worst outcome is a tie.' },
-    ],
-  },
-  {
-    heading: 'Lie and course conditions',
-    terms: [
-      { name: 'Tight lie', body: 'A ball sitting on bare or closely-mown ground with very little grass underneath. Demands clean contact.' },
-      { name: 'Plugged lie / Fried egg', body: 'Ball embedded in its own pitch mark, often in a bunker — only the top is visible, surrounded by disturbed sand.' },
-      { name: 'Hardpan', body: 'Very firm, dry ground with little or no grass. Produces low running shots; demands precise contact.' },
-    ],
-  },
-  {
-    heading: 'Equipment terms',
-    terms: [
-      { name: 'Muscle back', body: 'Iron with weight in a solid mass behind the center of the face — a blade. Less forgiving than cavity backs but unmatched feedback when struck pure.' },
-      { name: 'Cavity back', body: 'Iron with weight redistributed to the perimeter, creating a cavity. More forgiving on off-center hits. Standard for most amateurs.' },
-      { name: 'Loft', body: 'Angle of the clubface relative to vertical. Higher loft = higher flight, less distance. Lower = lower flight, more distance.' },
-      { name: 'Shaft flex', body: 'How much the shaft bends during the swing. Ladies, Senior, Regular, Stiff, Extra Stiff. Wrong flex affects flight and distance.' },
-      { name: 'Lie angle', body: 'Angle between shaft and ground when soled at address. Wrong lie angle pushes shots consistently left or right even with a good swing.' },
-    ],
-  },
-  {
-    heading: 'Slang and culture',
-    terms: [
-      { name: 'Worm burner', body: 'A shot that never gets airborne, rolling along the ground. Usually a topped or thinned iron.' },
-      { name: 'Shank', body: 'A shot struck off the hosel that fires hard right for a right-handed player. One of the most dreaded misses in golf — some players will not say the word.' },
-      { name: 'Yips', body: 'Involuntary muscle twitches or nerves on short putts (or chips). More mental than physical. Has ended careers.' },
-      { name: 'Cabbage', body: 'Thick, deep rough where the ball sits down and is hard to advance. You are just trying to get out.' },
-      { name: 'Lip out', body: 'A putt that catches the edge of the cup but does not fall in. One of the most frustrating outcomes in golf.' },
-      { name: 'Bite', body: 'What golfers shout asking a ball to stop on the green. Also a verb: "I need this to bite."' },
-      { name: 'Barkie', body: 'Making par after your ball hits a tree. Counted as bonus points in some social games.' },
-      { name: 'Greenie', body: 'In a group game, awarded to the player who hits the green in regulation on a par 3 closest to the pin.' },
-    ],
-  },
-  {
-    heading: 'Rules terms worth knowing',
-    terms: [
-      { name: 'Relief', body: 'When the rules let you move your ball without penalty — cart path, temporary water, GUR, embedded ball.' },
-      { name: 'Penalty area', body: 'Modern rules term (since 2019) for what was a water hazard. Yellow = two relief options, red = three. You may now ground your club.' },
-      { name: 'Ground under repair (GUR)', body: 'Areas marked for maintenance. Free relief.' },
-      { name: 'Provisional ball', body: 'A second ball played when the first may be lost or out of bounds. Must be announced. Saves walking back to replay.' },
-      { name: 'Stroke and distance', body: 'Penalty for a lost ball or OB. One stroke AND back to where you played from. The most severe penalty in golf.' },
-    ],
-  },
-  {
-    heading: 'Historical terms',
-    intro: 'Golf has a longer continuous history than almost any other sport, and its vocabulary reflects that. Some hickory-era terms still surface in writing, course names, or older players using them affectionately.',
-    terms: [
-      { name: 'Brassie', body: 'Old name for a 2-wood — named for the brass plate on the sole.' },
-      { name: 'Spoon', body: 'Old name for a 3-wood — for the shallow, spoon-like face.' },
-      { name: 'Baffy', body: 'Old name for a 4-wood. Rarely heard today.' },
-      { name: 'Cleek', body: 'Low-lofted hickory iron, roughly a modern 1 or 2-iron. Also a narrow-faced iron used for distance from tight lies.' },
-      { name: 'Mashie', body: 'Roughly a modern 5-iron. "Mashie niblick" was between a mashie and niblick — about a 7-iron.' },
-      { name: 'Niblick', body: 'High-lofted hickory club for short approaches and trouble — roughly a 9-iron or wedge.' },
-      { name: 'Jigger', body: 'Utility iron with moderate loft, used for punch shots and running approaches. No clean modern equivalent.' },
-      { name: 'Stymie', body: 'Pre-1952 match-play situation where your opponent\'s ball blocked your line. Still used figuratively — "in a stymie" means stuck.' },
-      { name: 'Fore', body: 'The warning shout when a ball is heading toward other players. Origin disputed — possibly from "forecaddie," possibly military. Still essential.' },
-    ],
-  },
-]
-
-function Stat({ label, value }: { label: string; value: string }) {
-  return (
-    <View style={{ minWidth: 80 }}>
-      <Text style={{ ...KICKER, color: '#5C6356', fontSize: 9, marginBottom: 4 }}>
-        {label}
-      </Text>
-      <Text
-        style={{
-          color: '#1C211C',
-          fontSize: 17,
-          fontWeight: '500',
-          fontVariant: ['tabular-nums'],
-        }}
-      >
-        {value}
-      </Text>
-    </View>
-  )
-}
-
-function HowToPracticeArticle() {
-  return (
-    <View>
-      <Text style={{ ...KICKER, marginBottom: 10 }}>Improving your game · Draft</Text>
-      <Text style={TITLE}>How to practice.</Text>
-
-      <H3>The uncomfortable truth</H3>
-      <Para>
-        Most golfers practice in a way that feels productive but isn't. They
-        hit a bucket of balls, stripe a few drives, feel good, and leave.
-        Their handicap doesn't move.
-      </Para>
-      <Para>
-        This isn't laziness. It's that no one ever explained what practice
-        is actually for — or that there's a difference between hitting balls
-        and getting better.
-      </Para>
-      <Para>
-        There's nothing wrong with going to the range to unwind, enjoy the
-        weather, or just swing a club. That's a legitimate and enjoyable
-        thing to do. But if your goal is to improve, that's a different
-        activity and it requires a different approach.
-      </Para>
-
-      <Divider />
-
-      <H3>What practice is actually for</H3>
-      <Para>
-        Practice exists to change your behavior on the golf course. Not to
-        feel good on the range. Not to hit your best shots. To change what
-        you do under pressure, on uneven lies, with something at stake.
-      </Para>
-      <Para>
-        If your practice doesn't eventually show up on the course, it wasn't
-        practice — it was entertainment. Which, again, is fine. Just know
-        which one you're doing.
-      </Para>
-
-      <Divider />
-
-      <H3>The four types of practice</H3>
-
-      <H4>Block practice</H4>
-      <KV label="What it is">
-        Repeating the same shot over and over. 50 7-irons in a row to the
-        same target.
-      </KV>
-      <KV label="Good for">
-        Learning a brand new movement. If you're working on a swing change
-        with your coach, some repetition helps establish the new pattern.
-        Use it sparingly and early.
-      </KV>
-      <KV label="Not good for">
-        Building skills that transfer to the course. Research consistently
-        shows that improvement during block practice doesn't stick. You get
-        better within the session and worse by the next round.
-      </KV>
-      <KV label="The trap">
-        Block practice feels like learning because you do get better within
-        the session. That feeling is misleading. It is one of the most
-        well-replicated findings in motor learning research.
-      </KV>
-      <Note variant="research">
-        Research basis: Robert Bjork (UCLA) — contextual interference
-        effect. See also: "Make It Stick" by Brown, Roediger, McDaniel.
-      </Note>
-      <Note variant="todo">TODO: Add specific study citations</Note>
-
-      <H4>Random practice</H4>
-      <KV label="What it is">
-        Varying every shot. 7-iron, driver, wedge, 5-iron — never hitting
-        the same club twice in a row.
-      </KV>
-      <KV label="Good for">
-        Building skills that actually transfer to the course. The research
-        on this is consistent across many studies. Random practice feels
-        harder and messier but produces dramatically better retention.
-      </KV>
-      <KV label="Why it works">
-        Each time you pick up a different club, your brain has to fully
-        reconstruct the motor pattern from scratch. That reconstruction
-        process is where learning happens.
-      </KV>
-      <KV label="Not good for">
-        Learning a brand new movement. Don't randomize before you have a
-        basic pattern to work with.
-      </KV>
-      <Note variant="research">
-        Research basis: Contextual interference effect, Battig (1979),
-        confirmed in many subsequent studies.
-      </Note>
-      <Note variant="todo">
-        TODO: Verify whether the beginner exception is well-established or
-        still debated
-      </Note>
-
-      <H4>Variable practice</H4>
-      <KV label="What it is">
-        Same club, different conditions. 9-iron from 100 yards, then 90,
-        then uphill, then into wind, then from a downslope.
-      </KV>
-      <KV label="Good for">
-        Building adaptability. Golf never gives you the same shot twice.
-        Variable practice trains you for that reality.
-      </KV>
-      <KV label="Combine it with">
-        Random practice for maximum transfer.
-      </KV>
-
-      <H4>Pressure practice</H4>
-      <KV label="What it is">
-        Creating consequences in practice. Something is at stake on each
-        shot.
-      </KV>
-      <Bullets
-        items={[
-          'Make 5 putts in a row from 6 feet or start over from 0',
-          'Last ball in your bucket must hit a specific target',
-          'Play a simulated 9 holes on the range — pick targets, keep score',
-          'Clock drill: 12 balls around the hole at 3 feet (clock positions), make all 12 in a row or start over',
-        ]}
-      />
-      <KV label="Good for">
-        Training your nervous system to perform under pressure. If you've
-        never practiced with consequences, your body hasn't learned how to
-        handle them on the course.
-      </KV>
-      <Note variant="research">
-        Research basis: Dr. Sian Beilock — "Choke" (2010). Bob Rotella —
-        "Golf Is Not a Game of Perfect" (1995).
-      </Note>
-      <Note variant="todo">
-        TODO: Add more specific pressure practice examples from tour player
-        documented routines
-      </Note>
-
-      <Divider />
-
-      <H3>How to structure a session</H3>
-      <Para>
-        A well-structured session has a flow. Adjust based on your time and
-        current focus area.
-      </Para>
-      <SessionRows />
-      <Note variant="todo">
-        TODO: Review these time allocations with a teaching professional.
-        The short game / full swing split in particular (Dave Pelz suggests
-        ~60% short game) needs more exploration.
-      </Note>
-
-      <Divider />
-
-      <H3>Having a goal for the session</H3>
-      <Para>
-        Before you hit a single ball, decide what you're trying to
-        accomplish today.
-      </Para>
-      <Bullets
-        items={[
-          'Bad goal: Work on my irons.',
-          'Better goal: Work on my ball striking.',
-          'Good goal: Hit 7 of 10 approach shots within 20 yards of the target from 150 yards.',
-        ]}
-      />
-      <Para>
-        The difference is measurability. If you can't tell whether you
-        achieved your goal, it wasn't a goal — it was a vague intention.
-      </Para>
-      <Para>
-        Your OGA strokes gained data tells you exactly where to focus. If
-        you're losing 1.2 strokes per round on approach shots, that's your
-        focus area. Your practice goal should be specific to that weakness.
-      </Para>
-      <H4>How to quantify your practice</H4>
-      <Bullets
-        items={[
-          'Track your success rate on pressure games over time',
-          'Note which club and distance you practiced',
-          'Write down what you worked on and what changed',
-          'Connect practice focus to round data over time',
-        ]}
-      />
-
-      <Divider />
-
-      <H3>What bad practice looks like</H3>
-
-      <H4>The range bucket spiral</H4>
-      <Para>
-        Buy a large bucket. Hit wedges, feel good. Move to 7-iron, stripe a
-        few. Pull out the driver. Spend 45 minutes hitting drivers because
-        that's the most fun. Leave feeling like you worked hard. Nothing
-        about your game changed.
-      </Para>
-
-      <H4>Same shot, same target, same lie</H4>
-      <Para>
-        You struggle with your 6-iron so you hit 60 6-irons to the same
-        target from a flat lie. On the course you face a 6-iron from a
-        downslope with water left. These are completely different skills.
-      </Para>
-
-      <H4>Practicing your strengths</H4>
-      <Para>
-        You putt well so you skip the green. You hit your driver well so
-        you spend an hour on the tee line. Improvement comes from
-        addressing weaknesses, not reinforcing strengths.
-      </Para>
-
-      <H4>No target, no feedback</H4>
-      <Para>
-        Hitting shots without a specific target or any way to evaluate the
-        result is exercise, not practice. Useful and enjoyable — but not
-        improvement.
-      </Para>
-
-      <H4>Working on technique under pressure</H4>
-      <Para>
-        The course is for playing. The range is for working on technique.
-        Working on a swing change during a round rarely ends well.
-      </Para>
-
-      <Divider />
-
-      <H3>Practice round vs scoring round</H3>
-      <Para>
-        Two completely different activities. Mixing them is one of the most
-        common mistakes amateurs make.
-      </Para>
-      <KV label="Practice round mode">
-        Experiment. Try the risky shot. Play from a bad lie on purpose. Hit
-        two balls. Explore. Score doesn't matter — you're gathering
-        information.
-      </KV>
-      <KV label="Scoring round mode">
-        Full pre-shot routine on every shot. Full commitment. No mulligans.
-        Track everything. This is performance mode.
-      </KV>
-      <Para>
-        The mistake is being half-committed to both — sort of practicing,
-        sort of scoring. You get the anxiety of performance without the
-        data of tracking, and the experimentation of practice without the
-        freedom of no consequences.
-      </Para>
-      <Para>
-        Before you tee off, decide which mode you're in and commit to it
-        fully.
-      </Para>
-
-      <Divider />
-
-      <H3>Resources</H3>
-      <Note variant="todo">
-        These are starting points for further research, not endorsements.
-        Verify all links are current before publishing.
-      </Note>
-
-      <H4>Books</H4>
-      {BOOKS.map((b) => (
-        <View
-          key={b.title}
-          style={{
-            borderTopWidth: 1,
-            borderColor: '#D9D2BF',
-            paddingVertical: 12,
-          }}
-        >
-          <Text
-            style={{
-              color: '#1C211C',
-              fontSize: 15,
-              fontStyle: 'italic',
-              fontWeight: '500',
-            }}
-          >
-            {b.title}
-            {b.by && (
-              <Text style={{ color: '#5C6356', fontStyle: 'normal', fontWeight: '400' }}>
-                {' '}— {b.by}
-              </Text>
-            )}
-          </Text>
-          <Text style={{ color: '#5C6356', fontSize: 14, lineHeight: 20, marginTop: 4 }}>
-            {b.note}
-          </Text>
-        </View>
-      ))}
-
-      <H4>Research worth knowing</H4>
-      <Bullets
-        items={[
-          'Robert Bjork — contextual interference, desirable difficulties (UCLA)',
-          'Anders Ericsson — deliberate practice and expert performance',
-          'Gabriele Wulf — attentional focus research showing external focus cues outperform internal focus cues',
-          'Sian Beilock — choking under pressure',
-        ]}
-      />
-
-      <H4>Online resources</H4>
-      <ResourceLink label="TPI (Titleist Performance Institute)" url="https://mytpi.com" />
-      <ResourceLink label="Robert Bjork's lab" url="https://bjorklab.psych.ucla.edu" />
-      <Note variant="todo" inline>
-        TODO: Verify Bjork lab URL is current
-      </Note>
-
-      <View
-        style={{
-          borderTopWidth: 1,
-          borderColor: '#D9D2BF',
-          paddingTop: 18,
-          marginTop: 22,
-        }}
-      >
-        <Text
-          style={{
-            ...KICKER,
-            color: '#8A8B7E',
-            lineHeight: 14,
-          }}
-        >
-          Last reviewed May 2026 · Draft, needs instructor review · Edit
-          docs/learn/how-to-practice.md to contribute
-        </Text>
-      </View>
-    </View>
-  )
-}
-
-function H3({ children }: { children: string }) {
-  return (
-    <Text
-      style={{
-        color: '#1C211C',
-        fontSize: 22,
-        fontStyle: 'italic',
-        fontWeight: '500',
-        lineHeight: 28,
-        marginTop: 20,
-        marginBottom: 12,
-      }}
-    >
-      {children}
-    </Text>
-  )
-}
-
-function H4({ children }: { children: string }) {
-  return (
-    <Text
-      style={{
-        color: '#1C211C',
-        fontSize: 17,
-        fontStyle: 'italic',
-        fontWeight: '500',
-        marginTop: 16,
-        marginBottom: 6,
-      }}
-    >
-      {children}
-    </Text>
-  )
-}
-
-function Para({ children }: { children: React.ReactNode }) {
-  return (
-    <Text
-      style={{
-        color: '#1C211C',
-        fontSize: 15,
-        lineHeight: 22,
-        marginBottom: 12,
-      }}
-    >
-      {children}
-    </Text>
-  )
-}
-
-function KV({ label, children }: { label: string; children: React.ReactNode }) {
-  return (
-    <Text
-      style={{
-        color: '#1C211C',
-        fontSize: 15,
-        lineHeight: 22,
-        marginBottom: 10,
-      }}
-    >
-      <Text
-        style={{
-          fontStyle: 'italic',
-          fontWeight: '500',
-          color: '#1C211C',
-        }}
-      >
-        {label}.{' '}
-      </Text>
-      {children}
-    </Text>
-  )
-}
-
-function Bullets({ items }: { items: string[] }) {
-  return (
-    <View style={{ marginBottom: 12 }}>
-      {items.map((item) => (
-        <View
-          key={item}
-          style={{
-            flexDirection: 'row',
-            marginBottom: 6,
-          }}
-        >
-          <Text style={{ color: '#1C211C', fontSize: 15, lineHeight: 22, marginRight: 6 }}>
-            •
-          </Text>
-          <Text
-            style={{
-              color: '#1C211C',
-              fontSize: 15,
-              lineHeight: 22,
-              flex: 1,
-            }}
-          >
-            {item}
-          </Text>
-        </View>
-      ))}
-    </View>
-  )
-}
-
-function Divider() {
-  return (
-    <View
-      style={{
-        borderTopWidth: 1,
-        borderColor: '#D9D2BF',
-        marginVertical: 18,
-      }}
-    />
-  )
-}
-
-function SessionRows() {
-  const rows = [
-    { phase: 'Warm up', dur: '10–15 min', why: 'Easy wedges, get body ready. Not practice time.' },
-    { phase: 'Skill work', dur: '20–30 min', why: 'Current focus area. Deliberate, with feedback.' },
-    { phase: 'Random / variable', dur: '20–30 min', why: 'Whole bag, random clubs, real targets.' },
-    { phase: 'Pressure games', dur: '10–15 min', why: 'Consequences on every shot. Keep score.' },
-    { phase: 'Short game', dur: '15–20 min', why: 'Chipping and pitching. Don’t skip.' },
-    { phase: 'Putting', dur: '10–15 min', why: 'Always end on the green. End by holing putts.' },
-  ]
-  return (
-    <View style={{ borderTopWidth: 1, borderColor: '#D9D2BF', marginBottom: 14 }}>
-      {rows.map((r) => (
-        <View
-          key={r.phase}
-          style={{
-            paddingVertical: 12,
-            borderBottomWidth: 1,
-            borderColor: '#D9D2BF',
-          }}
-        >
-          <View style={{ flexDirection: 'row', justifyContent: 'space-between', marginBottom: 4 }}>
-            <Text style={{ color: '#1C211C', fontSize: 15, fontStyle: 'italic', fontWeight: '500' }}>
-              {r.phase}
-            </Text>
-            <Text style={{ color: '#5C6356', fontSize: 12, fontVariant: ['tabular-nums'] }}>
-              {r.dur}
-            </Text>
-          </View>
-          <Text style={{ color: '#5C6356', fontSize: 14, lineHeight: 20 }}>{r.why}</Text>
-        </View>
-      ))}
-    </View>
-  )
-}
-
-function Note({
-  variant,
-  inline,
-  children,
-}: {
-  variant: 'research' | 'todo'
-  inline?: boolean
-  children: React.ReactNode
-}) {
-  if (!__DEV__) return null
-  const tone = variant === 'research' ? '#1F3D2C' : '#A66A1F'
-  const label = variant === 'research' ? 'Source' : 'Todo'
-  if (inline) {
-    return (
-      <Text style={{ ...KICKER, color: tone, marginTop: 4 }}>
-        [{label}] {children}
-      </Text>
-    )
-  }
-  return (
-    <View
-      style={{
-        backgroundColor: '#EBE5D6',
-        borderLeftWidth: 3,
-        borderLeftColor: tone,
-        padding: 12,
-        marginBottom: 12,
-      }}
-    >
-      <Text style={{ ...KICKER, color: tone, marginBottom: 4 }}>
-        {label} · dev only
-      </Text>
-      <Text style={{ color: '#5C6356', fontSize: 13, fontStyle: 'italic', lineHeight: 19 }}>
-        {children}
-      </Text>
-    </View>
-  )
-}
-
-function ResourceLink({ label, url }: { label: string; url: string }) {
-  return (
-    <Pressable onPress={() => Linking.openURL(url)} style={{ paddingVertical: 8 }}>
-      <Text style={{ color: '#1C211C', fontSize: 15, lineHeight: 22 }}>
-        {label}:{' '}
-        <Text style={{ color: '#1F3D2C', textDecorationLine: 'underline' }}>{url}</Text>
-      </Text>
-    </Pressable>
-  )
-}
-
-const BOOKS = [
-  {
-    title: '"Make It Stick"',
-    by: 'Brown, Roediger, McDaniel.',
-    note: 'Best plain-language summary of learning science. Not golf-specific but directly applicable.',
-  },
-  {
-    title: '"Golf Is Not a Game of Perfect"',
-    by: 'Bob Rotella.',
-    note: 'Standard text on golf psychology and performance.',
-  },
-  {
-    title: '"Dave Pelz\'s Short Game Bible"',
-    by: '',
-    note: 'Research-based approach to practice from inside 100 yards.',
-  },
-  {
-    title: '"Harvey Penick\'s Little Red Book"',
-    by: '',
-    note: 'The most beloved golf instruction book ever written. Simple, wise, feel-based.',
-  },
-  {
-    title: '"Peak"',
-    by: 'Anders Ericsson.',
-    note: 'His own account of deliberate practice research. Better than the Gladwell version.',
-  },
-  {
-    title: '"Choke"',
-    by: 'Sian Beilock.',
-    note: 'Accessible neuroscience of performance under pressure.',
-  },
-]
-
-const BENCHMARKS = [
-  {
-    label: 'Avg score',
-    pga: '69.5',
-    scratch: '72',
-    mid: '82 at 10 hcp',
-    high: '92 at 20 hcp',
-  },
-  { label: 'GIR %', pga: '67%', scratch: '50%', mid: '30%', high: '15%' },
-  { label: 'Putts / round', pga: '29', scratch: '32', mid: '34', high: '36' },
-  {
-    label: 'Driving distance',
-    pga: '294 yd',
-    scratch: '250 yd',
-    mid: '220 yd',
-    high: '190 yd',
-  },
-]
-
-function CourseManagementArticle() {
-  return (
-    <View>
-      <Text style={{ ...KICKER, marginBottom: 10 }}>On the course · Draft</Text>
-      <Text style={TITLE}>Course management.</Text>
-
-      <H3>You are not on the range anymore</H3>
-      <Para>
-        The driving range is where you work on your golf swing. The
-        golf course is where you work on your score. These are two
-        different activities requiring two different mindsets.
-      </Para>
-      <Para>
-        On the range, perfect is the goal. On the course, useful is
-        the goal. The player who shoots 78 with ugly but functional
-        golf beats the player who shoots 84 hitting it beautifully.
-        Score is the only thing the card cares about.
-      </Para>
-      <Para>
-        The moment you step onto the first tee, stop being a golfer
-        working on your swing and become a player trying to shoot
-        the lowest number possible with the tools you have today.
-        Your clubs are tools to get the ball in the hole. Not every
-        shot needs to be pretty. It just needs to work.
-      </Para>
-
-      <Divider />
-
-      <H3>Know your game, not your best game</H3>
-      <Para>
-        Course management starts with brutal honesty about what you
-        actually do — not what you're capable of on a perfect day.
-      </Para>
-      <Para>
-        Every player has tendencies. You probably miss in a
-        predictable direction. You probably have a club you don't
-        fully trust. You probably have a shot shape that shows up
-        under pressure whether you want it or not.
-      </Para>
-      <Para>
-        The player who knows they hit a soft fade under pressure and
-        plays for it will outscore the player who aims straight and
-        gets "surprised" by the same fade every time.
-      </Para>
-      <Para>Before you play, know:</Para>
-      <Bullets
-        items={[
-          'Your predominant miss direction',
-          "Which clubs you genuinely trust vs which ones you're hoping work out",
-          'How far you actually carry the ball — not your best carry, your reliable carry',
-          'Where your game breaks down under pressure',
-        ]}
-      />
-      <Para>
-        Your OGA strokes gained data tells you this more clearly
-        than your gut does. If SG approach is -1.4 per round, your
-        irons are leaking. If SG putting is +0.8, your putter is an
-        asset. Play accordingly.
-      </Para>
-
-      <Divider />
-
-      <H3>The confidence club rule</H3>
-      <Para>
-        For any shot on the course, choose the club you can hit the
-        shot you need 8 times out of 10. Not the club that gets you
-        there if you pure it. Not the club that's technically correct
-        on paper. The club you trust.
-      </Para>
-      <Para>
-        Standing over a shot with doubt is one of the most expensive
-        things you can do in golf. A committed swing with the "wrong"
-        club almost always beats a tentative swing with the "right"
-        one. If you're between clubs and one of them makes you
-        nervous — hit the other one. Every time.
-      </Para>
-      <Para>
-        Off the tee: hit the longest club you can confidently keep
-        in play. That is it. You don't need a perfect drive. You
-        need a ball in play. A 220-yard drive in the fairway beats a
-        280-yard drive in the trees every single time. Driver is not
-        always the answer and there is no shame in hitting 3-wood or
-        hybrid off the tee on a tight hole. The Playa doesn't care
-        what club anyone else is hitting. The Playa cares about the
-        score.
-      </Para>
-
-      <Divider />
-
-      <H3>The Way of the Playa</H3>
-      <Para>
-        Golf content creator Golf Sidekick has articulated a
-        philosophy called the Way of the Playa that every amateur
-        should understand. The core idea is simple: your ego is the
-        most expensive thing in your bag.
-      </Para>
-      <Para>
-        The Playa doesn't care about looking good. The Playa doesn't
-        care what club other people are hitting or whether the shot
-        looks impressive. The Playa cares about one thing — getting
-        the ball in the hole in as few strokes as possible — and
-        makes every decision in service of that goal.
-      </Para>
-      <Para>In practice this looks like:</Para>
-      <Bullets
-        items={[
-          'Hitting 3-wood off the tee when driver puts you in trouble',
-          'Laying up short of a hazard even when you think you can probably carry it',
-          'Chipping out sideways without shame when trees are in the way',
-          'Playing to the fat part of the green instead of attacking a tucked pin',
-          'Accepting the penalty and moving on instead of compounding the mistake',
-        ]}
-      />
-      <Para>
-        None of these decisions feel exciting. All of them save
-        strokes.
-      </Para>
-      <Para>
-        The Playa understands that golf is not a contest of who hits
-        the most impressive shots. It's a contest of who takes the
-        fewest. These are different games and most amateurs are
-        playing the wrong one.
-      </Para>
-      <Note variant="research">
-        Golf Sidekick — YouTube. Search "Way of the Playa." Practical,
-        ego-free approach to scoring.
-      </Note>
-
-      <Divider />
-
-      <H3>Aim at the center of the green</H3>
-      <Para>
-        This is a simple rule that will immediately save you strokes.
-      </Para>
-      <Para>
-        Most amateur golfers aim at the pin. The pin is rarely in
-        the center of the green. This means most amateur golfers are
-        constantly playing the hardest target available — the one
-        with the least margin for error.
-      </Para>
-      <Para>
-        Aim at the center of the green unless you have a very
-        specific, very good reason not to. The center of the green
-        is always a good shot. It never leaves you with a chip from
-        the wrong side. It gives you the most margin for your miss.
-      </Para>
-      <Para>
-        A birdie putt from 25 feet is still a birdie putt. What you
-        won't have is a chip from the rough, or a three-putt from 60
-        feet, or a false front situation because you were attacking
-        a back left pin from 170 yards.
-      </Para>
-
-      <Divider />
-
-      <H3>Plan the hole backwards</H3>
-      <Para>
-        Most amateurs walk up to the tee and hit their driver as far
-        as they can, then figure out what to do next. This is
-        reactive golf. It works well enough on straightforward holes
-        and falls apart on anything with a hazard, a difficult
-        green, or a premium on position.
-      </Para>
-      <Para>
-        The better approach is to plan the hole in reverse —
-        starting at the pin and working back to the tee.
-      </Para>
-      <Para>
-        Start on the green. Where is the pin today? Is it tucked
-        behind a bunker, on a shelf, near a false front? Now ask:
-        where on the green do I want to be putting from? That tells
-        you the ideal approach angle and the ideal landing zone for
-        your approach shot.
-      </Para>
-      <Para>
-        Now work back one more shot. To hit the approach from that
-        ideal spot, where does your previous shot need to land?
-        What's the right side of the fairway? What distance leaves
-        you a comfortable full shot rather than an awkward partial?
-      </Para>
-      <Para>
-        Now you know where your tee shot needs to go. Not just "in
-        the fairway" — a specific zone that sets up everything that
-        follows.
-      </Para>
-      <Para>
-        On a par 5: the same logic applies across three shots
-        instead of two. Where do you want to be for your pitch or
-        third shot? What does the layup need to accomplish to get
-        you there? Where does the tee shot need to go to make the
-        layup straightforward?
-      </Para>
-      <Para>
-        This is how caddies think. This is how tour professionals
-        think. And it's available to any amateur willing to spend 60
-        seconds on the tee thinking before swinging.
-      </Para>
-      <Para>
-        The practice round is where you build this map. Walk the
-        hole from green to tee. Stand where the approach shot will
-        be played from and look back at the fairway — you'll see
-        angles and landing zones you never noticed from the tee.
-        That knowledge compounds over time. Players who know a
-        course well aren't lucky. They've done this work.
-      </Para>
-
-      <Divider />
-
-      <H3>Your free throw range</H3>
-      <Para>
-        Every player has a distance where they feel genuinely
-        comfortable — a yardage where they know, without much doubt,
-        that they can get the ball close. This is your free throw
-        range.
-      </Para>
-      <Para>
-        For most amateurs it's somewhere between 50 and 100 yards.
-        For better players it might be 100-120. Whatever yours is,
-        identify it honestly.
-      </Para>
-      <Para>
-        Now start playing shots to leave yourself that number rather
-        than just maximizing distance on every shot. Course
-        management is partly about engineering the right approach
-        shot.
-      </Para>
-      <Para>
-        A 60-yard full wedge from a clean lie is easier than an
-        85-yard partial wedge from an awkward distance. Partial
-        wedge shots are hard. Full swing wedges are repeatable.
-        Don't manufacture difficult shots when strategy can avoid
-        them.
-      </Para>
-      <Para>
-        Off the tee on a long par 4: consider what club leaves you a
-        full shot in your free throw range rather than just
-        grip-and-ripping driver. On a par 5: a layup to your number
-        beats going for it from 230 yards with a marginal lie.
-      </Para>
-
-      <Divider />
-
-      <H3>The Scoring Method and the Scoring Zone</H3>
-      <Para>
-        Will Robins, PGA member and Golf Digest Best Young Teachers
-        honoree, developed a course management system called The
-        Scoring Method that reframes how amateurs think about every
-        hole.
-      </Para>
-      <Para>
-        The core concept is the Scoring Zone — a distance close
-        enough to the green that you're confident you can finish the
-        hole in 3 more shots or fewer. For most beginners and high
-        handicappers, start at 75 yards. As your game improves,
-        tighten it.
-      </Para>
-      <Para>
-        The Scoring Method tracks just two things per hole on a
-        modified scorecard:
-      </Para>
-      <H4>1. Did you reach the Scoring Zone in two shots?</H4>
-      <Para>
-        After your first two shots, are you inside your distance?
-        Yes or no.
-      </Para>
-      <H4>
-        2. Did you get up and down in 3 shots or fewer from the
-        Scoring Zone?
-      </H4>
-      <Para>
-        Once inside your zone, did you convert? Or did you take 4
-        from there?
-      </Para>
-      <Para>
-        Track these two numbers for a few rounds and patterns emerge
-        fast. Most amateurs discover they're actually reaching the
-        Scoring Zone regularly — the problem is converting once they
-        get there. That tells you exactly where to practice.
-      </Para>
-      <Para>
-        The deeper power of this system is that it shifts your
-        measure of success away from score and toward process. A
-        player who reaches the Scoring Zone in two shots and
-        converts every time will shoot in the 80s almost regardless
-        of how their ball-striking looks. The system shows you what
-        actually matters hole by hole.
-      </Para>
-      <Note variant="research">
-        Will Robins — The Scoring Method. thescoringmethod.com and
-        YouTube @thescoringmethod
-      </Note>
-
-      <Divider />
-
-      <H3>Think target, not trouble</H3>
-      <Para>
-        When you stand over a shot thinking "don't hit it in the
-        water" — you are thinking about the water. Your brain
-        doesn't process the "don't" particularly well. You are
-        programming yourself to think about exactly what you want to
-        avoid.
-      </Para>
-      <Para>
-        Instead: pick a specific, positive target. Not "away from
-        the bunker" but "at that tree on the left edge of the
-        fairway." A specific target gives your brain and body
-        something to work toward rather than something to flee from.
-      </Para>
-      <Para>
-        Manu from The Upbeat Golfer talks extensively about this —
-        committing fully to a target and a process before stepping
-        into the ball, then letting go of the result. The shot is
-        decided before you address it. Doubt after you've committed
-        is the enemy. Indecision kills golf shots more reliably than
-        poor mechanics.
-      </Para>
-      <Note variant="research">
-        The Upbeat Golfer (Manu) — YouTube. Process-driven approach,
-        target commitment, playing without fear.
-      </Note>
-
-      <Divider />
-
-      <H3>Never make two mistakes in a row</H3>
-      <Para>
-        You will hit bad shots. Every player at every level hits bad
-        shots. A bad shot is not a crisis — it's golf.
-      </Para>
-      <Para>
-        What turns a bad shot into a big number is the decision that
-        follows it.
-      </Para>
-      <Para>
-        After a bad shot, your only job is to get back into
-        position. Not to make up for it. Not to be a hero. Not to
-        erase it. Just to get somewhere you can play a normal golf
-        shot from.
-      </Para>
-      <Para>Punch out. Accept the bogey. Move on.</Para>
-      <Para>
-        The double bogey that becomes a triple happens because the
-        player tried to thread the needle through the trees instead
-        of punching out sideways. The bogey that becomes a triple
-        happens because the player went for the green from an
-        impossible lie when laying up was obviously the right call.
-      </Para>
-      <Para>
-        A bogey is one over par. A triple is three over. The shots
-        themselves might be identical — the decisions are what
-        separate them. The Playa takes his medicine every time. The
-        Playa never follows a bad shot with a stupid shot.
-      </Para>
-
-      <Divider />
-
-      <H3>Be process-driven, not results-driven</H3>
-      <Para>
-        Most golfers think too little on the golf course. They step
-        up to the ball, swing, and react emotionally to wherever it
-        goes. There's no plan and no process.
-      </Para>
-      <Para>
-        But there's an opposite problem: once golfers start trying
-        to improve, they often start thinking too much. They
-        outthink themselves — grinding over swing thoughts, worrying
-        about mechanics, replaying the last bad shot — because they
-        aren't yet good enough to execute what they're thinking
-        about. Analysis paralysis is just as damaging as
-        mindlessness.
-      </Para>
-      <Para>
-        The sweet spot is a clear, repeatable pre-shot process:
-      </Para>
-      <Bullets
-        items={[
-          'Pick a specific target',
-          'Commit fully to the shot',
-          'Go through your routine',
-          'Pull the trigger',
-        ]}
-      />
-      <Para>
-        That's it. You don't need to solve the hole. You need to
-        play one shot at a time with full commitment.
-      </Para>
-      <Para>
-        Try tracking — alongside your score — whether you committed
-        to each shot. Not whether it went where you wanted. Whether
-        you actually went through your process and pulled the
-        trigger without doubt.
-      </Para>
-      <Para>
-        Players who track this consistently find their scores drop
-        naturally. Not because they're grinding harder on results,
-        but because they've replaced reactive, emotional golf with a
-        repeatable approach. Results-oriented thinking leads to
-        tension, steering, and the exact outcome you were afraid of.
-        Process-oriented thinking gives you the best chance of
-        executing the shot.
-      </Para>
-
-      <Divider />
-
-      <H3>The practice round</H3>
-      <Para>
-        Course management is not something you can learn sitting in
-        a cart or watching videos. It's learned by doing — and the
-        practice round is where you do it.
-      </Para>
-      <Para>
-        Play two balls off the tee on every hole. Hit the aggressive
-        line you want to hit and the conservative line you think you
-        should hit. See which one actually leaves you the better
-        approach. You will be surprised how often the conservative
-        play gives you just as good an angle — sometimes better.
-      </Para>
-      <Para>
-        From trouble, play a second ball both ways. Hit the hero
-        shot you were tempted to play AND the safe punch-out. See
-        what actually happens. Most of the time the hero shot costs
-        you more than the punch-out. Your gut will stop suggesting
-        it so often once your eyes have seen the real results.
-      </Para>
-      <Para>
-        Use the practice round to find your Scoring Zone entry
-        points on each hole. Where do you need to be after two shots
-        to have a comfortable third? Work backwards from there to
-        plan your tee shot.
-      </Para>
-      <Para>
-        A practice round played thoughtfully and with intention is
-        worth more for course management than ten range sessions.
-        The course is the classroom.
-      </Para>
-
-      <Divider />
-
-      <H3>Keep it simple</H3>
-      <Para>
-        The fundamentals that will lower your score, in order of
-        importance:
-      </Para>
-      <Bullets
-        items={[
-          'Ball in play off the tee with the longest club you trust',
-          'Plan the hole backwards from the pin before you swing',
-          "Think about your target, not about where you can't go",
-          'Aim at the center of the green',
-          'Play to your free throw range',
-          'Commit to every shot through your full routine',
-          'Never follow a bad shot with a dumb shot',
-          'Take your medicine, make your bogey, move on',
-        ]}
-      />
-      <Para>
-        That's course management for most amateurs. It's not
-        complicated. It's just hard to actually do when there's
-        water on the left and your playing partners are watching.
-      </Para>
-      <Para>
-        The Way of the Playa is not a swing philosophy. It's a
-        mindset. And it's available to every player at every level
-        starting on the very next round they play.
-      </Para>
-
-      <Divider />
-
-      <H3>Resources</H3>
-      <Note variant="todo">
-        Verify all links before publishing.
-      </Note>
-      {RESOURCES.map((r) => (
-        <View
-          key={r.title}
-          style={{
-            borderTopWidth: 1,
-            borderColor: '#D9D2BF',
-            paddingVertical: 12,
-          }}
-        >
-          <Text
-            style={{
-              color: '#1C211C',
-              fontSize: 15,
-              fontStyle: 'italic',
-              fontWeight: '500',
-            }}
-          >
-            {r.title}
-            {r.by && (
-              <Text style={{ color: '#5C6356', fontStyle: 'normal', fontWeight: '400' }}>
-                {' '}— {r.by}
-              </Text>
-            )}
-          </Text>
-          <Text style={{ color: '#5C6356', fontSize: 14, lineHeight: 20, marginTop: 4 }}>
-            {r.note}
-          </Text>
-        </View>
-      ))}
-
-      <View
-        style={{
-          borderTopWidth: 1,
-          borderColor: '#D9D2BF',
-          paddingTop: 18,
-          marginTop: 22,
-        }}
-      >
-        <Text style={{ ...KICKER, color: '#8A8B7E', lineHeight: 14 }}>
-          Last reviewed May 2026 · Draft, needs instructor review ·
-          Edit docs/learn/course-management.md to contribute
-        </Text>
-      </View>
-    </View>
-  )
-}
-
-const RESOURCES = [
-  {
-    title: 'Golf Sidekick',
-    by: '',
-    note: 'YouTube. Search "Way of the Playa." Ego-free, score-focused course management for amateurs.',
-  },
-  {
-    title: 'The Upbeat Golfer (Manu)',
-    by: '',
-    note: 'YouTube. Process-driven mental approach and target commitment.',
-  },
-  {
-    title: 'Will Robins — The Scoring Method',
-    by: '',
-    note: 'thescoringmethod.com and YouTube @thescoringmethod. The Scoring Zone framework and modified scorecard system.',
-  },
-  {
-    title: '"Golf Is Not a Game of Perfect"',
-    by: 'Bob Rotella.',
-    note: 'The standard text on playing with what you have that day.',
-  },
-]
-
-function MentalGameArticle() {
-  return (
-    <View>
-      <Text style={{ ...KICKER, marginBottom: 10 }}>On the course · Draft</Text>
-      <Text style={TITLE}>The mental game.</Text>
-
-      <H3>Golf is you versus you</H3>
-      <Para>
-        Every other sport has an opponent actively trying to stop
-        you. In golf, the course is static. The ball sits still and
-        waits. Nobody is trying to beat you except you.
-      </Para>
-      <Para>
-        Once you have a functional swing, the game becomes almost
-        entirely mental. Most amateurs spend 95% of their practice
-        on the physical game and almost none on the mental game.
-        The swing is largely solved. The mental game is the frontier.
-      </Para>
-      <Para>
-        Golf is also one of the most humbling experiences available.
-        It will find every weakness in your ego and expose it
-        repeatedly. Either reduce the ego or suffer. There is no
-        third option.
-      </Para>
-
-      <Divider />
-
-      <H3>One shot at a time</H3>
-      <Para>
-        Every touring professional eventually says the same thing:
-        one shot at a time. It is repeated so often because it is
-        literally true. You cannot replay the last shot. You cannot
-        hit the next one until this one is done. The only shot that
-        exists is the one in front of you.
-      </Para>
-      <Para>
-        The brain doesn't work this way naturally. It replays the
-        shot you just botched. It fast-forwards to the putt on 18.
-        It calculates score. All of this thinking happens at exactly
-        the moment you need to be doing one thing.
-      </Para>
-      <Para>
-        Bob Rotella describes this as refusing to be seduced by
-        results. At the 2008 Masters, Trevor Immelman did not look
-        at a single leaderboard until walking up 18. The routine
-        protected him all day. The moment he stepped outside it,
-        the pressure arrived.
-      </Para>
-      <Para>
-        The short memory is a skill, not a trait. Give yourself a
-        window to feel the emotion — Rotella calls it walking it
-        off — then drop it. The bad shot already happened. Carrying
-        it into the next shot is a choice.
-      </Para>
-
-      <Divider />
-
-      <H3>Sulking won't get you anything</H3>
-      <Para>
-        When Padraig Harrington won the 2007 British Open, he
-        knocked two balls into the water on the last hole and made
-        double bogey. He still won. He told Rotella afterward that
-        it never entered his mind he might blow it. His only
-        thought was getting the ball in the hole so he could make
-        the playoff.
-      </Para>
-      <Para>
-        Rotella is direct: sulking won't get you anything. Feeling
-        sorry, replaying the mistake — none of it changes what
-        happened, all of it damages what comes next.
-      </Para>
-
-      <Divider />
-
-      <H3>The productive delusion</H3>
-      <Para>
-        Rotella's first rule: believe you can win. Not hope. Not
-        think you have a chance. Believe.
-      </Para>
-      <Para>
-        The player on a putting heater approaches the next putt
-        believing it's going in. The player who has been missing
-        all afternoon thinks: I'm due. Both are irrational. But the
-        golfer who believes stands over the ball with a different
-        quality of attention — and that affects the outcome.
-      </Para>
-      <Para>
-        It is the deliberate construction of a mental state that
-        serves performance. Call it confidence, call it a productive
-        delusion. You are allowed to be delusional in your own
-        favor. In golf, it is the correct strategy.
-      </Para>
-
-      <Divider />
-
-      <H3>Visualization</H3>
-      <Para>
-        Before a shot, the mental game starts. Not on the backswing
-        — before you take the club out of the bag.
-      </Para>
-      <Para>
-        See the shot you want to hit. The flight, the trajectory,
-        the bounce and roll. Like a video, in real time, as if it
-        already happened. For putting: see the line, see the ball
-        rolling along it, see it drop.
-      </Para>
-      <Para>
-        This is not mysticism. The brain rehearses motor patterns
-        through visualization in ways that translate into physical
-        execution. Pick a target, visualize the shot, let it rip.
-        In that order, every time.
-      </Para>
-      <Note variant="research">
-        Research basis: mental imagery and motor performance. See
-        Guillot &amp; Collet (2008) on mental simulation of motor
-        actions.
-      </Note>
-
-      <Divider />
-
-      <H3>The routine as an anchor</H3>
-      <Para>
-        A consistent pre-shot routine prepares you physically and
-        gives you a repeatable process. But it has a third purpose
-        most golfers miss: the routine is an anchor to the present
-        moment.
-      </Para>
-      <Para>
-        Curtis Strange won the 1988 U.S. Open looking calm on TV
-        but with his heart pounding. He told Rotella he was working
-        his tail off just to stay in the present. The routine made
-        him look calm and kept him functional.
-      </Para>
-      <Para>
-        When the mind is scattered, going through the routine
-        brings you back. Not magic — executing familiar physical
-        actions requires present-moment attention.
-      </Para>
-      <Para>
-        Routine must be consistent. Same grip check, same practice
-        swing, same target look — every time. The bigger the
-        moment, the more important it is the routine does not change.
-      </Para>
-      <Note variant="research">
-        Manu, The Upbeat Golfer — YouTube. Target commitment and
-        pre-shot routine for process-driven play.
-      </Note>
-
-      <Divider />
-
-      <H3>Patience over aggression</H3>
-      <Para>
-        Rotella's fourth rule: every time you have the urge to make
-        an aggressive play, go with the more conservative one. You
-        will always be okay.
-      </Para>
-      <Para>
-        At the 1992 U.S. Open, Tom Kite shot even par in 35-mph
-        wind and won by two. Most players didn't break 80. He did
-        it by staying patient and letting others beat themselves.
-        Patience compounds across 18 holes.
-      </Para>
-
-      <Divider />
-
-      <H3>Ignore unsolicited swing advice</H3>
-      <Para>
-        Rotella tells of a player who made eight birdies in round
-        one, stopped by the putting green, got two unsolicited
-        comments on his setup, and was a mess by the next day.
-      </Para>
-      <Para>
-        Stop them before they speak. Their comments will creep in
-        at the moment you need to be thinking about your target.
-        File anything observed for the range later — or let it go.
-      </Para>
-
-      <Divider />
-
-      <H3>Calm is a superpower</H3>
-      <Para>
-        You cannot make good decisions when angry. You cannot swing
-        well when amped up or seething. The state of anger is the
-        opposite of the state required to hit a golf ball.
-      </Para>
-      <Para>
-        This is not about being emotionless. Feel it briefly,
-        privately, then drop it. The next shot needs your full
-        attention.
-      </Para>
-      <Para>
-        Hand on chest, feel your breath, look at something specific
-        in your environment — grounding techniques exist to pull
-        you out of your own mind and back to the course where the
-        actual game is being played.
-      </Para>
-
-      <Divider />
-
-      <H3>Grounding and the present moment</H3>
-      <Para>
-        When the mind spirals, the way back is through the body,
-        not through more thinking. Feel your feet on the ground.
-        Feel the grip in your hands. Look at the grass, the trees,
-        the sky.
-      </Para>
-      <Para>
-        The mind in a spiral is pulling you into the past or the
-        future. The body only exists in the present. Returning
-        attention to physical sensation is the fastest route back
-        to where the golf shot actually lives.
-      </Para>
-      <Para>
-        Golf becomes deeply Buddhist here — not attached to
-        outcomes, not grasping at results, returning again and
-        again to what is actually happening right now. Everything
-        else is noise.
-      </Para>
-
-      <Divider />
-
-      <H3>Find someone who believes in you</H3>
-      <Para>
-        Ben Hogan considered quitting early in his career. Valerie
-        Hogan wouldn't let him. Confidence in yourself is essential.
-        Having someone who sees what you can't yet see in yourself
-        — a spouse, a friend, a coach — compounds it.
-      </Para>
-      <Para>
-        For amateurs this might be a playing partner who knows your
-        game, or a coach who genuinely believes in your potential.
-        Anyone who reflects your capability back to you on the days
-        you can't see it. The mental game is not fought entirely
-        alone.
-      </Para>
-
-      <Divider />
-
-      <H3>Be process-driven, not results-driven</H3>
-      <Para>
-        Most golfers think too little — react emotionally to where
-        the ball goes. The opposite problem: thinking too much.
-        Grinding swing thoughts, replaying the last bad shot.
-        Analysis paralysis is just as damaging as mindlessness.
-      </Para>
-      <Para>The sweet spot is a clear, repeatable pre-shot process:</Para>
-      <Bullets
-        items={[
-          'Pick a specific target',
-          'Visualize the shot',
-          'Commit fully',
-          'Go through your routine',
-          'Pull the trigger',
-        ]}
-      />
-      <Para>
-        Try tracking — alongside your score — whether you committed
-        to each shot. Not whether the result was what you wanted.
-        Whether you actually went through your process without
-        doubt. Process-oriented thinking gives you the best chance
-        of executing.
-      </Para>
-
-      <Divider />
-
-      <H3>It never gets fully solved</H3>
-      <Para>
-        Tour pros — with decades of experience and sports
-        psychologists — still get in their own way. The goal is
-        not to eliminate the mental challenge. It is to recover
-        more quickly. The pro who has a double on 6 and is back to
-        full focus on 7 has the same challenge as the amateur who
-        takes four holes to recover. They just recover faster.
-      </Para>
-      <Para>That recovery time is what you are training.</Para>
-
-      <Divider />
-
-      <H3>Resources</H3>
-      <Note variant="todo">
-        Verify all links before publishing.
-      </Note>
-      {MENTAL_GAME_RESOURCES.map((r) => (
-        <View
-          key={r.title}
-          style={{
-            borderTopWidth: 1,
-            borderColor: '#D9D2BF',
-            paddingVertical: 12,
-          }}
-        >
-          <Text
-            style={{
-              color: '#1C211C',
-              fontSize: 15,
-              fontStyle: 'italic',
-              fontWeight: '500',
-            }}
-          >
-            {r.title}
-            {r.by && (
-              <Text style={{ color: '#5C6356', fontStyle: 'normal', fontWeight: '400' }}>
-                {' '}— {r.by}
-              </Text>
-            )}
-          </Text>
-          <Text style={{ color: '#5C6356', fontSize: 14, lineHeight: 20, marginTop: 4 }}>
-            {r.note}
-          </Text>
-        </View>
-      ))}
-
-      <View
-        style={{
-          borderTopWidth: 1,
-          borderColor: '#D9D2BF',
-          paddingTop: 18,
-          marginTop: 22,
-        }}
-      >
-        <Text style={{ ...KICKER, color: '#8A8B7E', lineHeight: 14 }}>
-          Last reviewed May 2026 · Draft, needs review · Edit
-          docs/learn/mental-game.md to contribute
-        </Text>
-      </View>
-    </View>
-  )
-}
-
-const MENTAL_GAME_RESOURCES = [
-  {
-    title: '"Golf Is Not a Game of Perfect"',
-    by: 'Bob Rotella.',
-    note: 'The gold standard in golf psychology. Essential reading.',
-  },
-  {
-    title: '"Golf Is a Game of Confidence"',
-    by: 'Bob Rotella.',
-    note: 'The follow-up, equally valuable.',
-  },
-  {
-    title: 'Bob Rotella — "My 10 Rules on Mental Fitness"',
-    by: '',
-    note: 'GolfWRX. The ten rules referenced throughout this article.',
-  },
-  {
-    title: '"Zen Golf"',
-    by: 'Joseph Parent.',
-    note: 'Buddhist-influenced approach to present-moment play.',
-  },
-  {
-    title: 'The Upbeat Golfer (Manu)',
-    by: '',
-    note: 'YouTube. Process-driven mental approach, target commitment, playing without fear.',
-  },
-  {
-    title: '"The Inner Game of Tennis"',
-    by: 'Tim Gallwey.',
-    note: 'Not golf-specific but the foundational text on getting out of your own way in sport.',
-  },
-  {
-    title: '"Choke"',
-    by: 'Sian Beilock.',
-    note: 'The neuroscience of why we perform poorly under pressure and what to do about it.',
-  },
-]
-
-function SkillGamesArticle() {
-  return (
-    <View>
-      <Text style={{ ...KICKER, marginBottom: 10 }}>Improving your game · Draft</Text>
-      <Text style={TITLE}>Skill games and pressure games.</Text>
-
-      <H3>Why games beat mindless repetition</H3>
-      <Para>
-        Hitting the same shot over and over until you run out of
-        balls is the default range session. It feels like practice.
-        It largely isn't.
-      </Para>
-      <Para>
-        Games change the dynamic. A target, a score, stakes. You're
-        tracking something. You either succeed or you don't. That
-        structure forces commitment on each shot — the same
-        commitment you need on the course.
-      </Para>
-      <Para>
-        End sessions with games. Technical work belongs earlier
-        when focus is fresh. Finish with games when you're shifting
-        from mechanics to performance.
-      </Para>
-
-      <Divider />
-
-      <H3>The pressure problem</H3>
-      <Para>
-        The fundamental challenge of practice: you know you have
-        another ball. On the course, you don't. That knowledge
-        removes consequence from every range swing. You never
-        practice the mental state you need when something is
-        riding on the shot.
-      </Para>
-      <Para>
-        Pressure games solve this. Create enough consequence that
-        missing would genuinely bother you. Not so much that you're
-        miserable. Just enough that each shot matters.
-      </Para>
-      <Para>
-        Stakes must be real but not ruinous. A candy bar, ten
-        pushups, coffee with a buddy. Too small and you don't
-        care. Too large and anxiety takes over.
-      </Para>
-      <Para>
-        The last few balls of any range bucket are a natural
-        pressure game. You're out of balls. Make them count.
-      </Para>
-
-      <Divider />
-
-      <H3>Putting games</H3>
-
-      <H4>The clock drill</H4>
-      <Para>
-        Place twelve balls around a hole at three feet — clock
-        positions. Make all twelve in a row. Miss and start over.
-      </Para>
-      <Para>
-        Deceptively brutal. Getting to eleven and then missing is
-        genuinely painful. That pain is the point — it simulates
-        the pressure of needing to make a putt with something on
-        the line.
-      </Para>
-      <Para>
-        Move to four feet, then five. Tour players do this from
-        six feet as a warm-up. Make all twelve from six and your
-        short putting is tour-level.
-      </Para>
-      <KV label="Variation">
-        Use one ball. Walk around the clock, replace after each
-        putt, make all twelve in a row. Harder — no second chances,
-        and the walk resets your mental state between putts.
-      </KV>
-
-      <H4>Three-station pressure drill</H4>
-      <Para>
-        Three feet, six feet, nine feet. Make ten putts at each
-        station before moving on. Miss more than one and start the
-        station over.
-      </Para>
-      <Para>
-        Escalating pressure. Three feet is achievable. Six with
-        the threat of restart matters. Nine, after completing the
-        first two — real pressure. Track total attempts to complete
-        all three across sessions.
-      </Para>
-
-      <H4>The putting circuit</H4>
-      <Para>
-        Pick nine holes on the practice green. Par 2 each — one
-        putt, two-putt is par, three-putt is bogey. Play it as a
-        proper round. Same nine holes every session so you can
-        track improvement.
-      </Para>
-      <KV label="Competitive version">
-        Head to head with a partner. Stroke or match play. Adds
-        pressure solo practice can't replicate.
-      </KV>
-
-      <H4>The gate drill</H4>
-      <Para>
-        Two tees just wider than your putter, a foot in front of
-        the ball on your line. Roll the ball through without
-        touching either tee.
-      </Para>
-      <Para>
-        Trains starting line. Most missed putts are missed before
-        the ball breaks. Start at six feet straight, then ten,
-        then a breaker. Exposes inconsistency in stroke path fast.
-      </Para>
-
-      <H4>Speed control — the ladder drill</H4>
-      <Para>
-        Five balls at ten feet. Hit the first putt — note where it
-        finishes. The second must finish on the opposite side of
-        the hole from the first. The third must finish between the
-        first two.
-      </Para>
-      <Para>
-        Move back five feet and repeat. Trains intentional pace
-        variation — directly transferable to lag putting and fast
-        greens.
-      </Para>
-
-      <Divider />
-
-      <H3>Chipping and short game games</H3>
-
-      <H4>Up and down challenge</H4>
-      <Para>
-        Five spots around the practice green at varying distances.
-        From each spot, chip and putt out. Sum strokes. Set a
-        target — maybe 12 — and try to beat it. Lower as you
-        improve.
-      </Para>
-      <KV label="Pressure version">
-        Five up-and-downs in a row. Miss one and the streak
-        restarts. Getting to four and chunking the next chip is
-        exactly the pressure the course creates.
-      </KV>
-
-      <H4>Three clubs, one ball per station</H4>
-      <Para>
-        Three distances — 20, 40, 60 yards. One ball from each.
-        No mulligan. The single-ball constraint is everything.
-      </Para>
-
-      <H4>Flag left or flag right</H4>
-      <Para>
-        At the range pick a flag. Before each shot, declare: left
-        or right. Within ten feet on the declared side = birdie.
-        Correct side, outside the window = par. Wrong side =
-        double bogey.
-      </Para>
-      <Para>
-        Teaches two things. First, controlling shot shape under
-        pressure — you have to commit before you swing. Second,
-        the cost of the short side. Score over ten shots; -4 or
-        better is good.
-      </Para>
-
-      <Divider />
-
-      <H3>Full swing games</H3>
-
-      <H4>Horse at the range</H4>
-      <Para>
-        Like the basketball game. Pick a target. Hit. Partner has
-        to match it. Miss the match and you get a letter. First
-        to H-O-R-S-E loses. Calling your shot before swinging is
-        the same skill the course demands.
-      </Para>
-
-      <H4>The imaginary fairway</H4>
-      <Para>
-        Two targets define a fairway, 25–30 yards wide. Ten tee
-        shots. Driver in the fairway = 3, fairway wood = 2, iron
-        = 1. -2 if you imagine a hazard side and find it.
-      </Para>
-      <Para>
-        Incentivizes driver — the most distance, the most risk.
-        Forces a real reward-vs-risk decision on every shot.
-      </Para>
-
-      <H4>Same distance, different clubs</H4>
-      <Para>
-        Pick a flag at 100 yards. Hit it with five different clubs.
-        9-iron, 8-iron, PW, GW, SW — each requires a different
-        swing length and speed.
-      </Para>
-      <Para>
-        Builds feel for partial shots. Teaches your actual
-        distances better than full swings — you're forced to
-        feel a 75% swing with each club.
-      </Para>
-
-      <H4>Three club challenge</H4>
-      <Para>
-        Simulated nine-hole round on the range with only three
-        clubs — driver, mid iron, wedge. Tee shot, approach, chip
-        — each to a different target. Forces creativity. You
-        learn to work the ball.
-      </Para>
-
-      <Divider />
-
-      <H3>The last ball rule</H3>
-      <Para>
-        Whatever you're practicing, save your last ball for a
-        specific challenge. Announce it before you hit:
-      </Para>
-      <Bullets
-        items={[
-          '"This drive has to find the fairway."',
-          '"This wedge has to finish within ten feet of the flag."',
-          '"This 7-iron has to carry the 150 marker."',
-        ]}
-      />
-      <Para>
-        One ball, no second chance. The session ends on this shot
-        — and you find out whether the work translates when
-        something is riding on it.
-      </Para>
-
-      <Divider />
-
-      <H3>Building your own pressure games</H3>
-      <Para>
-        Generic drills work. Customized stakes work better.
-      </Para>
-      <KV label="Stakes must be real but not ruinous">
-        Candy bar, coffee, pushups, bragging rights. Twenty dollars
-        or public embarrassment shuts down performance instead of
-        training it.
-      </KV>
-      <KV label="Streaks beat totals">
-        10 of 15 creates less pressure than 5 in a row. Streaks
-        carry the weight of every previous success — closer to
-        how the course feels.
-      </KV>
-      <KV label="One-and-done beats best-of">
-        A single attempt, no mulligans. The closer to the
-        one-chance nature of a real shot, the better.
-      </KV>
-      <KV label="Track scores over time">
-        A game without a scoreboard is a drill. Watching your
-        score improve across months is genuinely motivating.
-      </KV>
-
-      <Divider />
-
-      <H3>Resources</H3>
-      <Note variant="todo">
-        Verify all links before publishing.
-      </Note>
-      {SKILL_GAMES_RESOURCES.map((r) => (
-        <View
-          key={r.title}
-          style={{
-            borderTopWidth: 1,
-            borderColor: '#D9D2BF',
-            paddingVertical: 12,
-          }}
-        >
-          <Text
-            style={{
-              color: '#1C211C',
-              fontSize: 15,
-              fontStyle: 'italic',
-              fontWeight: '500',
-            }}
-          >
-            {r.title}
-            {r.by && (
-              <Text style={{ color: '#5C6356', fontStyle: 'normal', fontWeight: '400' }}>
-                {' '}— {r.by}
-              </Text>
-            )}
-          </Text>
-          <Text style={{ color: '#5C6356', fontSize: 14, lineHeight: 20, marginTop: 4 }}>
-            {r.note}
-          </Text>
-        </View>
-      ))}
-
-      <View
-        style={{
-          borderTopWidth: 1,
-          borderColor: '#D9D2BF',
-          paddingTop: 18,
-          marginTop: 22,
-        }}
-      >
-        <Text style={{ ...KICKER, color: '#8A8B7E', lineHeight: 14 }}>
-          Last reviewed May 2026 · Draft, needs review · Edit
-          docs/learn/skill-games-pressure-games.md to contribute
-        </Text>
-      </View>
-    </View>
-  )
-}
-
-const SKILL_GAMES_RESOURCES = [
-  {
-    title: 'Golf Digest — 15 Best Golf Practice Games',
-    by: '',
-    note: 'golfdigest.com. Good overview of range and putting green games.',
-  },
-  {
-    title: 'Practical Golf — 5 Games That Build Real Skills',
-    by: '',
-    note: 'practical-golf.com. Solid pressure games with scoring systems.',
-  },
-  {
-    title: '"Dave Pelz\'s Short Game Bible"',
-    by: '',
-    note: 'Research-based approach to short game practice. Specific drills backed by data.',
-  },
-  {
-    title: 'The Upbeat Golfer (Manu)',
-    by: '',
-    note: 'YouTube. Process under pressure, routine under stakes.',
-  },
-]
