+++ apps/mobile/app/(app)/round/new.tsx
@@ -28,6 +28,7 @@ import {
 import type { Database } from '@oga/supabase'
 import { supabase } from '../../../lib/supabase'
 import { useAuth } from '../../../hooks/useAuth'
+import { FONT, TYPE } from '../../../lib/typography'
 
 type CourseRow = Database['public']['Tables']['courses']['Row']
 type HoleInsert = Database['public']['Tables']['holes']['Insert']
@@ -38,6 +39,7 @@ const KICKER: import('react-native').TextStyle = {
   fontWeight: '500',
   letterSpacing: 1.4,
   textTransform: 'uppercase',
+  fontFamily: FONT.mono,
 }
 
 interface GpsState {
@@ -361,13 +363,13 @@ export default function NewRound() {
         {mode === 'past' ? 'Log past round' : 'Start live round'}
       </Text>
       <Text
-        style={{
+        style={[TYPE.serif, {
           color: '#1C211C',
           fontSize: 28,
           fontWeight: '500',
           marginBottom: 14,
           fontStyle: 'italic',
-        }}
+        }]}
       >
         {mode === 'past' ? 'Pick the course you played' : 'Pick a course to start'}
       </Text>
@@ -376,7 +378,7 @@ export default function NewRound() {
         value={query}
         onChangeText={setQuery}
         autoCapitalize="words"
-        style={{
+        style={[TYPE.body, {
           backgroundColor: '#FBF8F1',
           borderWidth: 1,
           borderColor: '#D9D2BF',
@@ -385,7 +387,7 @@ export default function NewRound() {
           paddingVertical: 12,
           fontSize: 15,
           marginBottom: 14,
-        }}
+        }]}
       />
 
       {searching && (
@@ -409,18 +411,18 @@ export default function NewRound() {
                 }}
               >
                 <Text
-                  style={{
+                  style={[TYPE.bodyBold, {
                     color: '#1C211C',
                     fontSize: 15,
                     fontWeight: '500',
-                  }}
+                  }]}
                 >
                   {c.name}
                 </Text>
                 {(() => {
                   const where = [c.city, c.state].filter((s) => !!s).join(', ')
                   return where ? (
-                    <Text style={{ color: '#5C6356', fontSize: 12, marginTop: 2 }}>
+                    <Text style={[TYPE.body, { color: '#5C6356', fontSize: 12, marginTop: 2 }]}>
                       {where}
                     </Text>
                   ) : null
@@ -446,16 +448,16 @@ export default function NewRound() {
                 }}
               >
                 <Text
-                  style={{
+                  style={[TYPE.bodyBold, {
                     color: '#1C211C',
                     fontSize: 15,
                     fontWeight: '500',
-                  }}
+                  }]}
                 >
                   {r.name}
                 </Text>
                 {formatLocation(r) ? (
-                  <Text style={{ color: '#5C6356', fontSize: 12, marginTop: 2 }}>
+                  <Text style={[TYPE.body, { color: '#5C6356', fontSize: 12, marginTop: 2 }]}>
                     {formatLocation(r)}
                   </Text>
                 ) : null}
@@ -477,12 +479,12 @@ export default function NewRound() {
             }}
           >
             <Text
-              style={{
+              style={[TYPE.bodyBold, {
                 color: '#1F3D2C',
                 fontSize: 14,
                 fontWeight: '600',
                 letterSpacing: 0.3,
-              }}
+              }]}
             >
               {hasResults ? "Don't see your course? Add it →" : 'Course not found? Add it →'}
             </Text>
@@ -490,7 +492,7 @@ export default function NewRound() {
         )}
 
         {error && (
-          <Text style={{ color: '#A33A2A', fontSize: 13, marginTop: 12 }}>
+          <Text style={[TYPE.body, { color: '#A33A2A', fontSize: 13, marginTop: 12 }]}>
             {error}
           </Text>
         )}
@@ -546,13 +548,13 @@ function ManualCourseForm({
       >
         <Text style={{ ...KICKER, marginBottom: 8 }}>Add course</Text>
         <Text
-          style={{
+          style={[TYPE.serif, {
             color: '#1C211C',
             fontSize: 28,
             fontStyle: 'italic',
             fontWeight: '500',
             marginBottom: 18,
-          }}
+          }]}
         >
           New course
         </Text>
@@ -629,12 +631,12 @@ function ManualCourseForm({
                 }}
               >
                 <Text
-                  style={{
+                  style={[TYPE.serifUpright, {
                     color: '#1C211C',
                     fontSize: 15,
                     fontWeight: '500',
                     fontVariant: ['tabular-nums'],
-                  }}
+                  }]}
                 >
                   {p}
                 </Text>
@@ -644,11 +646,11 @@ function ManualCourseForm({
         </View>
 
         <Text
-          style={{
+          style={[TYPE.body, {
             color: '#8A8B7E',
             fontSize: 12,
             marginTop: 18,
-          }}
+          }]}
         >
           {gpsCoords
             ? `GPS captured (${gpsCoords.lat.toFixed(4)}, ${gpsCoords.lng.toFixed(4)}) — set as hole 1 tee.`
@@ -667,7 +669,7 @@ function ManualCourseForm({
               alignItems: 'center',
             }}
           >
-            <Text style={{ color: '#5C6356', fontSize: 13 }}>Cancel</Text>
+            <Text style={[TYPE.body, { color: '#5C6356', fontSize: 13 }]}>Cancel</Text>
           </Pressable>
           <Pressable
             onPress={() =>
@@ -683,12 +685,12 @@ function ManualCourseForm({
             }}
           >
             <Text
-              style={{
+              style={[TYPE.bodyBold, {
                 color: '#F2EEE5',
                 fontSize: 14,
                 fontWeight: '600',
                 letterSpacing: 0.3,
-              }}
+              }]}
             >
               {busy ? 'Creating…' : 'Create course →'}
             </Text>
@@ -719,11 +721,11 @@ function Chip({
       }}
     >
       <Text
-        style={{
+        style={[active ? TYPE.bodyBold : TYPE.body, {
           color: active ? '#F2EEE5' : '#1C211C',
           fontSize: 13,
           fontWeight: active ? '600' : '400',
-        }}
+        }]}
       >
         {label}
       </Text>
@@ -740,4 +742,5 @@ const inputStyle = {
   paddingVertical: 12,
   fontSize: 15,
   color: '#1C211C',
+  fontFamily: FONT.body,
 } as const
