+++ apps/mobile/components/home/RoundCTAs.tsx
@@ -1,5 +1,6 @@
 import { Pressable, Text } from 'react-native'
 import { Link } from 'expo-router'
+import { TYPE } from '../../lib/typography'
 
 export function StartLiveRoundCTA() {
   return (
@@ -17,24 +18,30 @@ export function StartLiveRoundCTA() {
           }}
         >
           <Text
-            style={{
-              color: '#F2EEE5',
-              fontSize: 16,
-              fontWeight: '700',
-              letterSpacing: 0.4,
-            }}
+            style={[
+              TYPE.bodyBold,
+              {
+                color: '#F2EEE5',
+                fontSize: 16,
+                fontWeight: '700',
+                letterSpacing: 0.4,
+              },
+            ]}
           >
             ▶  Start live round
           </Text>
         </Pressable>
       </Link>
       <Text
-        style={{
-          color: '#5C6356',
-          fontSize: 12,
-          textAlign: 'center',
-          marginBottom: 14,
-        }}
+        style={[
+          TYPE.body,
+          {
+            color: '#5C6356',
+            fontSize: 12,
+            textAlign: 'center',
+            marginBottom: 14,
+          },
+        ]}
       >
         Track shots in real time with GPS
       </Text>
@@ -59,12 +66,15 @@ export function LogPastRoundCTA() {
         }}
       >
         <Text
-          style={{
-            color: '#1F3D2C',
-            fontSize: 14,
-            fontWeight: '600',
-            letterSpacing: 0.3,
-          }}
+          style={[
+            TYPE.bodyBold,
+            {
+              color: '#1F3D2C',
+              fontSize: 14,
+              fontWeight: '600',
+              letterSpacing: 0.3,
+            },
+          ]}
         >
           +  Log past round
         </Text>
