+++ apps/mobile/components/learn/articles/HowToPractice.tsx
@@ -1,0 +1,378 @@
+import { Linking, Pressable, Text, View } from 'react-native'
+import {
+  ArticleFooter,
+  ArticleHeader,
+  BulletList,
+  C,
+  DevNote,
+  H3,
+  H4,
+  Hr,
+  Kv,
+  P,
+  ResourceList,
+} from '../primitives'
+import { FONT } from '../../../lib/typography'
+
+export function HowToPracticeArticle() {
+  return (
+    <View>
+      <ArticleHeader kicker="Improving your game · Draft" title="How to practice." />
+
+      <H3>The uncomfortable truth</H3>
+      <P>
+        Most golfers practice in a way that feels productive but isn't. They
+        hit a bucket of balls, stripe a few drives, feel good, and leave.
+        Their handicap doesn't move.
+      </P>
+      <P>
+        This isn't laziness. It's that no one ever explained what practice
+        is actually for — or that there's a difference between hitting balls
+        and getting better.
+      </P>
+      <P>
+        There's nothing wrong with going to the range to unwind, enjoy the
+        weather, or just swing a club. That's a legitimate and enjoyable
+        thing to do. But if your goal is to improve, that's a different
+        activity and it requires a different approach.
+      </P>
+
+      <Hr />
+
+      <H3>What practice is actually for</H3>
+      <P>
+        Practice exists to change your behavior on the golf course. Not to
+        feel good on the range. Not to hit your best shots. To change what
+        you do under pressure, on uneven lies, with something at stake.
+      </P>
+      <P>
+        If your practice doesn't eventually show up on the course, it wasn't
+        practice — it was entertainment. Which, again, is fine. Just know
+        which one you're doing.
+      </P>
+
+      <Hr />
+
+      <H3>The four types of practice</H3>
+
+      <H4>Block practice</H4>
+      <Kv label="What it is">
+        Repeating the same shot over and over. 50 7-irons in a row to the
+        same target.
+      </Kv>
+      <Kv label="Good for">
+        Learning a brand new movement. If you're working on a swing change
+        with your coach, some repetition helps establish the new pattern.
+        Use it sparingly and early.
+      </Kv>
+      <Kv label="Not good for">
+        Building skills that transfer to the course. Research consistently
+        shows that improvement during block practice doesn't stick. You get
+        better within the session and worse by the next round.
+      </Kv>
+      <Kv label="The trap">
+        Block practice feels like learning because you do get better within
+        the session. That feeling is misleading. It is one of the most
+        well-replicated findings in motor learning research.
+      </Kv>
+      <DevNote variant="research">
+        Research basis: Robert Bjork (UCLA) — contextual interference
+        effect. See also: "Make It Stick" by Brown, Roediger, McDaniel.
+      </DevNote>
+      <DevNote variant="todo">TODO: Add specific study citations</DevNote>
+
+      <H4>Random practice</H4>
+      <Kv label="What it is">
+        Varying every shot. 7-iron, driver, wedge, 5-iron — never hitting
+        the same club twice in a row.
+      </Kv>
+      <Kv label="Good for">
+        Building skills that actually transfer to the course. The research
+        on this is consistent across many studies. Random practice feels
+        harder and messier but produces dramatically better retention.
+      </Kv>
+      <Kv label="Why it works">
+        Each time you pick up a different club, your brain has to fully
+        reconstruct the motor pattern from scratch. That reconstruction
+        process is where learning happens.
+      </Kv>
+      <Kv label="Not good for">
+        Learning a brand new movement. Don't randomize before you have a
+        basic pattern to work with.
+      </Kv>
+      <DevNote variant="research">
+        Research basis: Contextual interference effect, Battig (1979),
+        confirmed in many subsequent studies.
+      </DevNote>
+      <DevNote variant="todo">
+        TODO: Verify whether the beginner exception is well-established or
+        still debated
+      </DevNote>
+
+      <H4>Variable practice</H4>
+      <Kv label="What it is">
+        Same club, different conditions. 9-iron from 100 yards, then 90,
+        then uphill, then into wind, then from a downslope.
+      </Kv>
+      <Kv label="Good for">
+        Building adaptability. Golf never gives you the same shot twice.
+        Variable practice trains you for that reality.
+      </Kv>
+      <Kv label="Combine it with">
+        Random practice for maximum transfer.
+      </Kv>
+
+      <H4>Pressure practice</H4>
+      <Kv label="What it is">
+        Creating consequences in practice. Something is at stake on each
+        shot.
+      </Kv>
+      <BulletList
+        items={[
+          'Make 5 putts in a row from 6 feet or start over from 0',
+          'Last ball in your bucket must hit a specific target',
+          'Play a simulated 9 holes on the range — pick targets, keep score',
+          'Clock drill: 12 balls around the hole at 3 feet (clock positions), make all 12 in a row or start over',
+        ]}
+      />
+      <Kv label="Good for">
+        Training your nervous system to perform under pressure. If you've
+        never practiced with consequences, your body hasn't learned how to
+        handle them on the course.
+      </Kv>
+      <DevNote variant="research">
+        Research basis: Dr. Sian Beilock — "Choke" (2010). Bob Rotella —
+        "Golf Is Not a Game of Perfect" (1995).
+      </DevNote>
+      <DevNote variant="todo">
+        TODO: Add more specific pressure practice examples from tour player
+        documented routines
+      </DevNote>
+
+      <Hr />
+
+      <H3>How to structure a session</H3>
+      <P>
+        A well-structured session has a flow. Adjust based on your time and
+        current focus area.
+      </P>
+      <SessionRows />
+      <DevNote variant="todo">
+        TODO: Review these time allocations with a teaching professional.
+        The short game / full swing split in particular (Dave Pelz suggests
+        ~60% short game) needs more exploration.
+      </DevNote>
+
+      <Hr />
+
+      <H3>Having a goal for the session</H3>
+      <P>
+        Before you hit a single ball, decide what you're trying to
+        accomplish today.
+      </P>
+      <BulletList
+        items={[
+          'Bad goal: Work on my irons.',
+          'Better goal: Work on my ball striking.',
+          'Good goal: Hit 7 of 10 approach shots within 20 yards of the target from 150 yards.',
+        ]}
+      />
+      <P>
+        The difference is measurability. If you can't tell whether you
+        achieved your goal, it wasn't a goal — it was a vague intention.
+      </P>
+      <P>
+        Your OGA strokes gained data tells you exactly where to focus. If
+        you're losing 1.2 strokes per round on approach shots, that's your
+        focus area. Your practice goal should be specific to that weakness.
+      </P>
+      <H4>How to quantify your practice</H4>
+      <BulletList
+        items={[
+          'Track your success rate on pressure games over time',
+          'Note which club and distance you practiced',
+          'Write down what you worked on and what changed',
+          'Connect practice focus to round data over time',
+        ]}
+      />
+
+      <Hr />
+
+      <H3>What bad practice looks like</H3>
+
+      <H4>The range bucket spiral</H4>
+      <P>
+        Buy a large bucket. Hit wedges, feel good. Move to 7-iron, stripe a
+        few. Pull out the driver. Spend 45 minutes hitting drivers because
+        that's the most fun. Leave feeling like you worked hard. Nothing
+        about your game changed.
+      </P>
+
+      <H4>Same shot, same target, same lie</H4>
+      <P>
+        You struggle with your 6-iron so you hit 60 6-irons to the same
+        target from a flat lie. On the course you face a 6-iron from a
+        downslope with water left. These are completely different skills.
+      </P>
+
+      <H4>Practicing your strengths</H4>
+      <P>
+        You putt well so you skip the green. You hit your driver well so
+        you spend an hour on the tee line. Improvement comes from
+        addressing weaknesses, not reinforcing strengths.
+      </P>
+
+      <H4>No target, no feedback</H4>
+      <P>
+        Hitting shots without a specific target or any way to evaluate the
+        result is exercise, not practice. Useful and enjoyable — but not
+        improvement.
+      </P>
+
+      <H4>Working on technique under pressure</H4>
+      <P>
+        The course is for playing. The range is for working on technique.
+        Working on a swing change during a round rarely ends well.
+      </P>
+
+      <Hr />
+
+      <H3>Practice round vs scoring round</H3>
+      <P>
+        Two completely different activities. Mixing them is one of the most
+        common mistakes amateurs make.
+      </P>
+      <Kv label="Practice round mode">
+        Experiment. Try the risky shot. Play from a bad lie on purpose. Hit
+        two balls. Explore. Score doesn't matter — you're gathering
+        information.
+      </Kv>
+      <Kv label="Scoring round mode">
+        Full pre-shot routine on every shot. Full commitment. No mulligans.
+        Track everything. This is performance mode.
+      </Kv>
+      <P>
+        The mistake is being half-committed to both — sort of practicing,
+        sort of scoring. You get the anxiety of performance without the
+        data of tracking, and the experimentation of practice without the
+        freedom of no consequences.
+      </P>
+      <P>
+        Before you tee off, decide which mode you're in and commit to it
+        fully.
+      </P>
+
+      <Hr />
+
+      <H3>Resources</H3>
+      <DevNote variant="todo">
+        These are starting points for further research, not endorsements.
+        Verify all links are current before publishing.
+      </DevNote>
+
+      <H4>Books</H4>
+      <ResourceList items={BOOKS} />
+
+      <H4>Research worth knowing</H4>
+      <BulletList
+        items={[
+          'Robert Bjork — contextual interference, desirable difficulties (UCLA)',
+          'Anders Ericsson — deliberate practice and expert performance',
+          'Gabriele Wulf — attentional focus research showing external focus cues outperform internal focus cues',
+          'Sian Beilock — choking under pressure',
+        ]}
+      />
+
+      <H4>Online resources</H4>
+      <ResourceLink label="TPI (Titleist Performance Institute)" url="https://mytpi.com" />
+      <ResourceLink label="Robert Bjork's lab" url="https://bjorklab.psych.ucla.edu" />
+      <DevNote variant="todo" inline>
+        TODO: Verify Bjork lab URL is current
+      </DevNote>
+
+      <ArticleFooter>
+        Last reviewed May 2026 · Draft, needs instructor review · Edit
+        docs/learn/how-to-practice.md to contribute
+      </ArticleFooter>
+    </View>
+  )
+}
+
+function SessionRows() {
+  const rows = [
+    { phase: 'Warm up', dur: '10–15 min', why: 'Easy wedges, get body ready. Not practice time.' },
+    { phase: 'Skill work', dur: '20–30 min', why: 'Current focus area. Deliberate, with feedback.' },
+    { phase: 'Random / variable', dur: '20–30 min', why: 'Whole bag, random clubs, real targets.' },
+    { phase: 'Pressure games', dur: '10–15 min', why: 'Consequences on every shot. Keep score.' },
+    { phase: 'Short game', dur: '15–20 min', why: 'Chipping and pitching. Don’t skip.' },
+    { phase: 'Putting', dur: '10–15 min', why: 'Always end on the green. End by holing putts.' },
+  ]
+  return (
+    <View style={{ borderTopWidth: 1, borderColor: C.line, marginBottom: 14 }}>
+      {rows.map((r) => (
+        <View
+          key={r.phase}
+          style={{
+            paddingVertical: 12,
+            borderBottomWidth: 1,
+            borderColor: C.line,
+          }}
+        >
+          <View style={{ flexDirection: 'row', justifyContent: 'space-between', marginBottom: 4 }}>
+            <Text style={{ color: C.ink, fontFamily: FONT.serifItalic, fontSize: 15, fontStyle: 'italic', fontWeight: '500' }}>
+              {r.phase}
+            </Text>
+            <Text style={{ color: C.inkDim, fontFamily: FONT.mono, fontSize: 12, fontVariant: ['tabular-nums'] }}>
+              {r.dur}
+            </Text>
+          </View>
+          <Text style={{ color: C.inkDim, fontFamily: FONT.body, fontSize: 14, lineHeight: 20 }}>{r.why}</Text>
+        </View>
+      ))}
+    </View>
+  )
+}
+
+function ResourceLink({ label, url }: { label: string; url: string }) {
+  return (
+    <Pressable onPress={() => Linking.openURL(url)} style={{ paddingVertical: 8 }}>
+      <Text style={{ color: C.ink, fontFamily: FONT.body, fontSize: 15, lineHeight: 22 }}>
+        {label}:{' '}
+        <Text style={{ color: C.accent, textDecorationLine: 'underline' }}>{url}</Text>
+      </Text>
+    </Pressable>
+  )
+}
+
+const BOOKS = [
+  {
+    title: '"Make It Stick"',
+    by: 'Brown, Roediger, McDaniel.',
+    note: 'Best plain-language summary of learning science. Not golf-specific but directly applicable.',
+  },
+  {
+    title: '"Golf Is Not a Game of Perfect"',
+    by: 'Bob Rotella.',
+    note: 'Standard text on golf psychology and performance.',
+  },
+  {
+    title: '"Dave Pelz\'s Short Game Bible"',
+    by: '',
+    note: 'Research-based approach to practice from inside 100 yards.',
+  },
+  {
+    title: '"Harvey Penick\'s Little Red Book"',
+    by: '',
+    note: 'The most beloved golf instruction book ever written. Simple, wise, feel-based.',
+  },
+  {
+    title: '"Peak"',
+    by: 'Anders Ericsson.',
+    note: 'His own account of deliberate practice research. Better than the Gladwell version.',
+  },
+  {
+    title: '"Choke"',
+    by: 'Sian Beilock.',
+    note: 'Accessible neuroscience of performance under pressure.',
+  },
+]
