+++ apps/mobile/components/learn/articles/MentalGame.tsx
@@ -1,0 +1,340 @@
+import { View } from 'react-native'
+import {
+  ArticleFooter,
+  ArticleHeader,
+  BulletList,
+  DevNote,
+  H3,
+  Hr,
+  P,
+  ResourceList,
+} from '../primitives'
+
+export function MentalGameArticle() {
+  return (
+    <View>
+      <ArticleHeader kicker="On the course · Draft" title="The mental game." />
+
+      <H3>Golf is you versus you</H3>
+      <P>
+        Every other sport has an opponent actively trying to stop
+        you. In golf, the course is static. The ball sits still and
+        waits. Nobody is trying to beat you except you.
+      </P>
+      <P>
+        Once you have a functional swing, the game becomes almost
+        entirely mental. Most amateurs spend 95% of their practice
+        on the physical game and almost none on the mental game.
+        The swing is largely solved. The mental game is the frontier.
+      </P>
+      <P>
+        Golf is also one of the most humbling experiences available.
+        It will find every weakness in your ego and expose it
+        repeatedly. Either reduce the ego or suffer. There is no
+        third option.
+      </P>
+
+      <Hr />
+
+      <H3>One shot at a time</H3>
+      <P>
+        Every touring professional eventually says the same thing:
+        one shot at a time. It is repeated so often because it is
+        literally true. You cannot replay the last shot. You cannot
+        hit the next one until this one is done. The only shot that
+        exists is the one in front of you.
+      </P>
+      <P>
+        The brain doesn't work this way naturally. It replays the
+        shot you just botched. It fast-forwards to the putt on 18.
+        It calculates score. All of this thinking happens at exactly
+        the moment you need to be doing one thing.
+      </P>
+      <P>
+        Bob Rotella describes this as refusing to be seduced by
+        results. At the 2008 Masters, Trevor Immelman did not look
+        at a single leaderboard until walking up 18. The routine
+        protected him all day. The moment he stepped outside it,
+        the pressure arrived.
+      </P>
+      <P>
+        The short memory is a skill, not a trait. Give yourself a
+        window to feel the emotion — Rotella calls it walking it
+        off — then drop it. The bad shot already happened. Carrying
+        it into the next shot is a choice.
+      </P>
+
+      <Hr />
+
+      <H3>Sulking won't get you anything</H3>
+      <P>
+        When Padraig Harrington won the 2007 British Open, he
+        knocked two balls into the water on the last hole and made
+        double bogey. He still won. He told Rotella afterward that
+        it never entered his mind he might blow it. His only
+        thought was getting the ball in the hole so he could make
+        the playoff.
+      </P>
+      <P>
+        Rotella is direct: sulking won't get you anything. Feeling
+        sorry, replaying the mistake — none of it changes what
+        happened, all of it damages what comes next.
+      </P>
+
+      <Hr />
+
+      <H3>The productive delusion</H3>
+      <P>
+        Rotella's first rule: believe you can win. Not hope. Not
+        think you have a chance. Believe.
+      </P>
+      <P>
+        The player on a putting heater approaches the next putt
+        believing it's going in. The player who has been missing
+        all afternoon thinks: I'm due. Both are irrational. But the
+        golfer who believes stands over the ball with a different
+        quality of attention — and that affects the outcome.
+      </P>
+      <P>
+        It is the deliberate construction of a mental state that
+        serves performance. Call it confidence, call it a productive
+        delusion. You are allowed to be delusional in your own
+        favor. In golf, it is the correct strategy.
+      </P>
+
+      <Hr />
+
+      <H3>Visualization</H3>
+      <P>
+        Before a shot, the mental game starts. Not on the backswing
+        — before you take the club out of the bag.
+      </P>
+      <P>
+        See the shot you want to hit. The flight, the trajectory,
+        the bounce and roll. Like a video, in real time, as if it
+        already happened. For putting: see the line, see the ball
+        rolling along it, see it drop.
+      </P>
+      <P>
+        This is not mysticism. The brain rehearses motor patterns
+        through visualization in ways that translate into physical
+        execution. Pick a target, visualize the shot, let it rip.
+        In that order, every time.
+      </P>
+      <DevNote variant="research">
+        Research basis: mental imagery and motor performance. See
+        Guillot &amp; Collet (2008) on mental simulation of motor
+        actions.
+      </DevNote>
+
+      <Hr />
+
+      <H3>The routine as an anchor</H3>
+      <P>
+        A consistent pre-shot routine prepares you physically and
+        gives you a repeatable process. But it has a third purpose
+        most golfers miss: the routine is an anchor to the present
+        moment.
+      </P>
+      <P>
+        Curtis Strange won the 1988 U.S. Open looking calm on TV
+        but with his heart pounding. He told Rotella he was working
+        his tail off just to stay in the present. The routine made
+        him look calm and kept him functional.
+      </P>
+      <P>
+        When the mind is scattered, going through the routine
+        brings you back. Not magic — executing familiar physical
+        actions requires present-moment attention.
+      </P>
+      <P>
+        Routine must be consistent. Same grip check, same practice
+        swing, same target look — every time. The bigger the
+        moment, the more important it is the routine does not change.
+      </P>
+      <DevNote variant="research">
+        Manu, The Upbeat Golfer — YouTube. Target commitment and
+        pre-shot routine for process-driven play.
+      </DevNote>
+
+      <Hr />
+
+      <H3>Patience over aggression</H3>
+      <P>
+        Rotella's fourth rule: every time you have the urge to make
+        an aggressive play, go with the more conservative one. You
+        will always be okay.
+      </P>
+      <P>
+        At the 1992 U.S. Open, Tom Kite shot even par in 35-mph
+        wind and won by two. Most players didn't break 80. He did
+        it by staying patient and letting others beat themselves.
+        Patience compounds across 18 holes.
+      </P>
+
+      <Hr />
+
+      <H3>Ignore unsolicited swing advice</H3>
+      <P>
+        Rotella tells of a player who made eight birdies in round
+        one, stopped by the putting green, got two unsolicited
+        comments on his setup, and was a mess by the next day.
+      </P>
+      <P>
+        Stop them before they speak. Their comments will creep in
+        at the moment you need to be thinking about your target.
+        File anything observed for the range later — or let it go.
+      </P>
+
+      <Hr />
+
+      <H3>Calm is a superpower</H3>
+      <P>
+        You cannot make good decisions when angry. You cannot swing
+        well when amped up or seething. The state of anger is the
+        opposite of the state required to hit a golf ball.
+      </P>
+      <P>
+        This is not about being emotionless. Feel it briefly,
+        privately, then drop it. The next shot needs your full
+        attention.
+      </P>
+      <P>
+        Hand on chest, feel your breath, look at something specific
+        in your environment — grounding techniques exist to pull
+        you out of your own mind and back to the course where the
+        actual game is being played.
+      </P>
+
+      <Hr />
+
+      <H3>Grounding and the present moment</H3>
+      <P>
+        When the mind spirals, the way back is through the body,
+        not through more thinking. Feel your feet on the ground.
+        Feel the grip in your hands. Look at the grass, the trees,
+        the sky.
+      </P>
+      <P>
+        The mind in a spiral is pulling you into the past or the
+        future. The body only exists in the present. Returning
+        attention to physical sensation is the fastest route back
+        to where the golf shot actually lives.
+      </P>
+      <P>
+        Golf becomes deeply Buddhist here — not attached to
+        outcomes, not grasping at results, returning again and
+        again to what is actually happening right now. Everything
+        else is noise.
+      </P>
+
+      <Hr />
+
+      <H3>Find someone who believes in you</H3>
+      <P>
+        Ben Hogan considered quitting early in his career. Valerie
+        Hogan wouldn't let him. Confidence in yourself is essential.
+        Having someone who sees what you can't yet see in yourself
+        — a spouse, a friend, a coach — compounds it.
+      </P>
+      <P>
+        For amateurs this might be a playing partner who knows your
+        game, or a coach who genuinely believes in your potential.
+        Anyone who reflects your capability back to you on the days
+        you can't see it. The mental game is not fought entirely
+        alone.
+      </P>
+
+      <Hr />
+
+      <H3>Be process-driven, not results-driven</H3>
+      <P>
+        Most golfers think too little — react emotionally to where
+        the ball goes. The opposite problem: thinking too much.
+        Grinding swing thoughts, replaying the last bad shot.
+        Analysis paralysis is just as damaging as mindlessness.
+      </P>
+      <P>The sweet spot is a clear, repeatable pre-shot process:</P>
+      <BulletList
+        items={[
+          'Pick a specific target',
+          'Visualize the shot',
+          'Commit fully',
+          'Go through your routine',
+          'Pull the trigger',
+        ]}
+      />
+      <P>
+        Try tracking — alongside your score — whether you committed
+        to each shot. Not whether the result was what you wanted.
+        Whether you actually went through your process without
+        doubt. Process-oriented thinking gives you the best chance
+        of executing.
+      </P>
+
+      <Hr />
+
+      <H3>It never gets fully solved</H3>
+      <P>
+        Tour pros — with decades of experience and sports
+        psychologists — still get in their own way. The goal is
+        not to eliminate the mental challenge. It is to recover
+        more quickly. The pro who has a double on 6 and is back to
+        full focus on 7 has the same challenge as the amateur who
+        takes four holes to recover. They just recover faster.
+      </P>
+      <P>That recovery time is what you are training.</P>
+
+      <Hr />
+
+      <H3>Resources</H3>
+      <DevNote variant="todo">
+        Verify all links before publishing.
+      </DevNote>
+      <ResourceList items={MENTAL_GAME_RESOURCES} />
+
+      <ArticleFooter>
+        Last reviewed May 2026 · Draft, needs review · Edit
+        docs/learn/mental-game.md to contribute
+      </ArticleFooter>
+    </View>
+  )
+}
+
+const MENTAL_GAME_RESOURCES = [
+  {
+    title: '"Golf Is Not a Game of Perfect"',
+    by: 'Bob Rotella.',
+    note: 'The gold standard in golf psychology. Essential reading.',
+  },
+  {
+    title: '"Golf Is a Game of Confidence"',
+    by: 'Bob Rotella.',
+    note: 'The follow-up, equally valuable.',
+  },
+  {
+    title: 'Bob Rotella — "My 10 Rules on Mental Fitness"',
+    by: '',
+    note: 'GolfWRX. The ten rules referenced throughout this article.',
+  },
+  {
+    title: '"Zen Golf"',
+    by: 'Joseph Parent.',
+    note: 'Buddhist-influenced approach to present-moment play.',
+  },
+  {
+    title: 'The Upbeat Golfer (Manu)',
+    by: '',
+    note: 'YouTube. Process-driven mental approach, target commitment, playing without fear.',
+  },
+  {
+    title: '"The Inner Game of Tennis"',
+    by: 'Tim Gallwey.',
+    note: 'Not golf-specific but the foundational text on getting out of your own way in sport.',
+  },
+  {
+    title: '"Choke"',
+    by: 'Sian Beilock.',
+    note: 'The neuroscience of why we perform poorly under pressure and what to do about it.',
+  },
+]
